% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{add_DQ}
\alias{add_DQ}
\title{Add DQ information to the processed data}
\usage{
add_DQ(proc, curr_batch)
}
\arguments{
\item{proc}{A data frame containing the processed anomaly or signal data, which must include a `GPS` column.}

\item{curr_batch}{A `ts` object with `dqmask` and `dq.lev` attributes, representing the DQ mask and level.}
}
\value{
A `data.frame` with additional columns corresponding to DQ flags.
If \code{dq.lev = "all"}, all available DQ channels are appended. Otherwise, only the specified level is added.
}
\description{
Adds data quality (DQ) mask information to a `proc` data frame, based on the
DQ information stored in the current batch (`curr_batch`).
The function aligns DQ values by flooring the GPS timestamp and performs sampling
rate adjustments if necessary.
}
