% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{add_P0_DQ}
\alias{add_P0_DQ}
\title{Add P0_DQ mask to processed data}
\usage{
add_P0_DQ(proc, DQ = "BURST_CAT2")
}
\arguments{
\item{proc}{A data frame containing the processed anomaly data with a \code{P0} column and one or more DQ columns.}

\item{DQ}{A character vector of one or more DQ channel names. Defaults to \code{"BURST_CAT2"}.}
}
\value{
A `data.frame` with added column(s) \code{P0_<DQ>} that contain \code{P0} values masked by duty cycle flags.
}
\description{
Adds a masked version of the \code{P0} column (e.g., false alarm probability)
based on the duty cycle specified by the \code{DQ} flags. If a DQ flag indicates
downtime (\code{DQ == 0}), the corresponding \code{P0} value is replaced with \code{NA}.
}
