% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{add_Pexp}
\alias{add_Pexp}
\title{Add exponential inter-cluster probability}
\usage{
add_Pexp(proc, updated_stat)
}
\arguments{
\item{proc}{A dataframe from the anomaly pipeline containing `t_lag`.}

\item{updated_stat}{A list containing `stats$lambda_c`, typically returned from `update_stat()` or `update_logic()`.}
}
\value{
The input dataframe with an added `Pexp` column.
}
\description{
Computes and adds an exponential CDF-based column `Pexp` to the anomaly dataframe,
representing the probability \eqn{P(t \le T)} for each cluster's inter-cluster gap \eqn{t_{\rm lag}}.
}
\details{
This is calculated as \code{pexp(t_lag, lambda_c)} where \code{lambda_c} is the cluster rate
estimated from \code{updated_stat}.
}
