% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{add_Ppois}
\alias{add_Ppois}
\title{Add Poisson anomaly-tail probability to dataframe}
\usage{
add_Ppois(proc, updated_stat)
}
\arguments{
\item{proc}{A dataframe from the pipeline output containing `N_anom` per cluster.}

\item{updated_stat}{A list containing `stats$lambda_a`, typically returned from `update_stat()` or `update_logic()`.}
}
\value{
The input dataframe `proc` with an added `Ppois` column.
}
\description{
Computes and appends a new column `Ppois` to the processed anomaly dataframe,
representing the Poisson-tail probability \eqn{P(n \ge N_{\rm anom})}
using the current estimated rate \eqn{\lambda_a} from `updated_stat`.
}
