% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{amax}
\alias{amax}
\title{Calculate maximum of absolute values}
\usage{
amax(x, na.rm = T)
}
\arguments{
\item{na.rm}{A logical (default: TRUE). Whether remove NA values or not.}

\item{xA}{numeric vector.}
}
\value{
A numeric.
}
\description{
Calculate maximum of absolute values
}
