% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{amin}
\alias{amin}
\title{Calculate minimum of absolute values}
\usage{
amin(x, na.rm = T)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical (default: TRUE). Whether to remove NA values.}
}
\value{
A numeric. The minimum of absolute values.
}
\description{
Calculate minimum of absolute values
}
