% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{append.line}
\alias{append.line}
\title{Append a Line to an Existing File}
\usage{
append.line(x, file, show = T)
}
\arguments{
\item{x}{A character or object to write.}

\item{file}{A character string specifying the file path.}

\item{show}{Logical; if \code{TRUE}, print the file content (default: TRUE).}
}
\value{
None.
}
\description{
Append content to a file and optionally print it.
}
