% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{apply_pca}
\alias{apply_pca}
\title{Perform PCA with fixed sign for interpretability}
\usage{
apply_pca(x, retx = T, center = F, scale. = F, tol = NULL, rank. = NULL, ...)
}
\arguments{
\item{x}{A data frame or matrix.}

\item{retx, center, scale., tol, rank., ...}{Passed to `prcomp()`.}
}
\value{
A modified PCA object with signs adjusted.
}
\description{
Perform PCA with fixed sign for interpretability
}
