% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{ar.resid}
\alias{ar.resid}
\title{Simple AR wrapper}
\usage{
ar.resid(ts, ...)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{...}{Additional arguments for 'ar'.}
}
\value{
A list with `$res: ar(ts, ...)$resid` and `$AR: ar(ts, ...)` result itself.
}
\description{
Simple AR wrapper
}
