% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{ari_mean}
\alias{ari_mean}
\title{Arithmetic Mean Ignoring NA}
\usage{
ari_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric scalar representing the mean.
}
\description{
Compute the arithmetic mean of a numeric vector, ignoring \code{NA} values.
}
