% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{as_tbt}
\alias{as_tbt}
\title{Convert `ts` to tibbletime object}
\usage{
as_tbt(x, time_col = "time", val_col = "x")
}
\arguments{
\item{x}{A `ts` object.}

\item{time_col}{A character. Column name for time (default: "time").}

\item{val_col}{A character. Column name for value (default: "x").}
}
\value{
A \code{tbl_time} object.
}
\description{
Convert `ts` to tibbletime object
}
