% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{auto.diff}
\alias{auto.diff}
\title{Automatic difference by KPSS test}
\usage{
auto.diff(ts, t.seg = 0.5, d_max = 2, verbose = TRUE)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{t.seg}{A numeric (default: 0.5). Time length of each segment in second.}
}
\value{
A list with `$d`: the optimal order of difference, `$out`: A differenced time series (`ts`) object, and `$p.values`: A list containing data frames of p-values for KPSS test result.
}
\description{
Automatic difference by KPSS test
}
