% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{Autoregressive}
\alias{Autoregressive}
\title{Fit AR model using either single or ensemble approach}
\usage{
Autoregressive(x, p, aic = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{x}{A `ts` object.}

\item{p}{AR order(s).}

\item{aic}{Use AIC.}

\item{verbose}{Verbose output.}

\item{...}{Additional args.}
}
\value{
Residual `ts` object with AR metadata.
}
\description{
Fit AR model using either single or ensemble approach
}
