% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{BandPass}
\alias{BandPass}
\title{Apply band-pass filter to `ts`}
\usage{
BandPass(x, fl = NULL, fu = NULL, resp = "FIR", filt_order = NULL, verbose = T)
}
\arguments{
\item{x}{A `ts` object.}

\item{fl}{A numeric. Lower cutoff frequency.}

\item{fu}{A numeric. Upper cutoff frequency.}

\item{resp}{A character. Filter type ("FIR" or "IIR").}

\item{filt_order}{A numeric. Filter order (default: 512 for FIR, 8 for IIR).}

\item{verbose}{Logical. If \code{TRUE}, print filter info.}
}
\value{
A band-pass filtered `ts` object with attributes.
}
\description{
Apply band-pass filter to `ts`
}
