% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{batching.network}
\alias{batching.network}
\title{Perform batching and NA filtering for multiple ts objects}
\usage{
batching.network(ts.list, t_bch = 1, has.DQ = TRUE)
}
\arguments{
\item{ts.list}{A named list of `ts` objects, one per detector}

\item{t_bch}{Batch length in seconds (default: 1)}

\item{has.DQ}{Logical; whether to use DQ mask in batching (default: TRUE)}
}
\value{
A reshaped list of valid (non-NaN) batches from all detectors
}
\description{
Perform batching and NA filtering for multiple ts objects
}
