% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{C.Burg}
\alias{C.Burg}
\title{AR Burg core (.Call)}
\usage{
C.Burg(x, order.max)
}
\arguments{
\item{x}{numeric vector (time series)}

\item{order.max}{integer, maximum AR order}
}
\value{
list(coefs = p x p numeric matrix, var.pred1 = numeric(p+1), var.pred2 = numeric(p+1))
}
\description{
AR Burg core (.Call)
}
