% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{C_Burg}
\alias{C_Burg}
\title{AR Burg core (C interface)}
\usage{
C_Burg(x, order.max)
}
\arguments{
\item{x}{Input series.}

\item{order.max}{Maximum AR order.}
}
\value{
List with AR coefficient matrix and prediction variances.
}
\description{
AR Burg core (C interface)
}
