% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{CCF}
\alias{CCF}
\title{Calculate Cross Correlation Function and plot it}
\usage{
CCF(
  ts1,
  ts2,
  lag.max = length(ts1),
  na.action = na.fail,
  title = NULL,
  plot = TRUE,
  save = FALSE,
  dir = NULL,
  prefix = NULL,
  times = NULL
)
}
\description{
Calculate Cross Correlation Function and plot it
}
