% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{ceiling_digit}
\alias{ceiling_digit}
\title{Ceiling with Decimal Digits}
\usage{
ceiling_digit(x, digits = 0)
}
\arguments{
\item{x}{A numeric value or vector.}

\item{digits}{Integer (default: 0). Number of decimal digits to retain.}
}
\value{
A numeric vector with values ceiled to specified digits.
}
\description{
Applies the \code{ceiling()} function to a number with precision controlled by number of digits.
}
