% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{check.installed}
\alias{check.installed}
\title{Check installation}
\usage{
check.installed(packages, install = TRUE)
}
\arguments{
\item{packages}{A character or a vector. Package name(s).}

\item{install}{A logical (default: TRUE). Whether it automatically install those packages that are not installed yet.}
}
\description{
Check installation
}
