% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{check_normality}
\alias{check_normality}
\title{Check stationary with AD test}
\usage{
check_normality(ts, t.seg = 0.5)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{t.seg}{A numeric (default: 0.5). Time length of each segment in second.}
}
\value{
A numeric vector of p.values.
}
\description{
Check stationary with AD test
}
