% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{check_stationary}
\alias{check_stationary}
\title{Check stationary with KPSS test}
\usage{
check_stationary(ts, t.seg = 0.5)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{t.seg}{A numeric (default: 0.5). Time length of each segment in second.}
}
\value{
A data frame of p.values for both level and trend.
}
\description{
Check stationary with KPSS test
}
