% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{concat_ts}
\alias{concat_ts}
\title{Concatenate with previous tail to form new ts}
\usage{
concat_ts(prev = NA, curr, n_former)
}
\arguments{
\item{prev}{A `ts` object or `NA` (default). Previous time series.}

\item{curr}{A `ts` object. Current time series.}

\item{n_former}{An integer. Number of tail points from `prev` to keep.}
}
\value{
A new concatenated `ts` object.
}
\description{
Concatenates the tail of the previous time series (`prev`) with the current one (`curr`),
keeping only the last `n_former` points from `prev`. If `prev` is `NA`, returns `curr` unchanged.
}
\examples{
prev <- ts(1:100, start = 0, frequency = 1)
curr <- ts(101:105, start = 100, frequency = 1)
concat_ts(prev, curr, n_former = 5)

}
