% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{correlate}
\alias{correlate}
\title{Correlate of two complex vectors}
\usage{
correlate(x, y)
}
\description{
Reference: https://pycbc.org/pycbc/latest/html/_modules/pycbc/filter/matchedfilter_numpy.html#correlate
}
\details{
Python \code{correlate} function example
\code{
import numpy
def correlate(x, y, z):
     z.data[:] = numpy.conjugate(x.data)[:]
     z *= y
}
}
