% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{create_wd}
\alias{create_wd}
\title{Create WD recursively with alternative option}
\usage{
create_wd(root, sub, subsub, alternative = F, set = T)
}
\arguments{
\item{root}{A character. A root directory (full path).}

\item{sub}{A character. A sub-directory (only name).}

\item{subsub}{A character. A sub-sub-directory (only name).}

\item{alternative}{A logical (default: FALSE). If directory is already exists, it puts suffix, e.g. _1, _2 ...}

\item{set}{A logical (default: TRUE). Whether automatically set the working directory after creating it.}
}
\value{
A character. A final working directory path, if 'set=TRUE' (for any use for later).
}
\description{
Create WD recursively with alternative option
}
