% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{crop_to}
\alias{crop_to}
\title{Handy window function with index}
\usage{
crop_to(ts, ind.range)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{ind.range}{A vector of length of 2. A first and second element refer to initial and final index, respectively.}
}
\value{
A cropped time series (`ts`) object.
}
\description{
Handy window function with index
}
