% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{cutoff_to}
\alias{cutoff_to}
\title{Crop frequency series to a given range}
\usage{
cutoff_to(x, ref, frange = NULL)
}
\arguments{
\item{x}{A `fs` object (frequency series).}

\item{ref}{A vector with elements: `c(min, max)`, indicating index range.}

\item{frange}{A vector of frequency range. `frange[1]` can be larger than `ref[1]` and `frange[2]` can be smaller than `ref[2]`, so as to consider the specific frequency range.}
}
\value{
A cropped `fs` object with preserved attributes.
}
\description{
Crop frequency series to a given range
}
