% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{decomp_freq_trend}
\alias{decomp_freq_trend}
\title{Estimate decomposition windows for frequency and trend components}
\usage{
decomp_freq_trend(
  ts,
  fac.f = 1,
  fac.t = 1.5,
  alpha = 0.05,
  max_period_frac = 0.2,
  lag.max = NULL,
  use_fft_fallback = TRUE
)
}
\arguments{
\item{ts}{A \code{ts} object. Input time series.}

\item{fac.f}{A numeric multiplier (default: 2) for the frequency window length.}

\item{fac.t}{A numeric multiplier (default: 10) for the trend window length.}
}
\description{
Heuristically determines suitable window lengths for frequency and trend decomposition
based on the autocorrelation function (ACF) of a time series.
}
