% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{deltaf}
\alias{deltaf}
\title{Extract Frequency Resolution from `fs` Object}
\usage{
deltaf(x)
}
\arguments{
\item{x}{An `fs` object.}
}
\value{
A numeric scalar representing frequency resolution.
}
\description{
Returns the `delta_f` attribute from an `fs` object.
}
