% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QT.R
\name{deltam_f}
\alias{deltam_f}
\title{Convert mismatch to fractional spacing}
\usage{
deltam_f(mismatch)
}
\arguments{
\item{mismatch}{A numeric. Desired fractional mismatch (e.g., 0.2).}
}
\value{
A numeric. Fractional spacing.
}
\description{
Compute the fractional spacing (delta m) between neighboring tiles from mismatch.
}
