% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{Differencing}
\alias{Differencing}
\title{Difference a time series manually or automatically}
\usage{
Differencing(ts, d, t.seg = 0.5, return.pvals = FALSE, verbose = TRUE)
}
\arguments{
\item{ts}{A `ts` object.}

\item{d}{Differencing order. Use 0 for none, integer for fixed, 'auto' for automatic.}

\item{t.seg}{Segment duration.}

\item{return.pvals}{Return KPSS p-values.}

\item{verbose}{Whether to print messages.}
}
\value{
A differenced `ts` object with metadata.
}
\description{
Difference a time series manually or automatically
}
