% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTrasnform.R
\name{.iter_freqs}
\alias{.iter_freqs}
\title{Iterate over the frequencies of this 'Qplane'}
\usage{
.iter_freqs(q, frange, mismatch, duration)
}
\arguments{
\item{q}{A numeric. q value.}

\item{frange}{A numeric vector. upper and lower bounds of frequency range.}

\item{mismatch}{A numeric. percentage of desired fractional mismatch.}

\item{dur}{A numeric. duration of timeseries in seconds.}
}
\value{
A numeric vector. Q-Tile frequency.
}
\description{
Iterate over the frequencies of this 'Qplane'
}
