% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QT.R
\name{.iter_qs}
\alias{.iter_qs}
\title{Iterate over the Q values}
\usage{
.iter_qs(qrange, deltam)
}
\arguments{
\item{qrange}{A numeric vector. upper and lower bounds of q range.}

\item{deltam}{A numeric. Fractional mismatch between neighboring tiles.}
}
\value{
A numeric. Q value for Q-tile
}
\description{
Iterate over the Q values
}
