% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSC.R
\name{download_gwosc}
\alias{download_gwosc}
\title{Download strain data from GWOSC}
\usage{
download_gwosc(
  destfile = "/tmp/tmp.hdf5",
  event_name,
  det,
  dur = 32,
  file.format = "hdf5",
  sampling.freq = 4096,
  version = "latest",
  timeout = 300,
  load = FALSE,
  remove = FALSE,
  direct.url = NULL
)
}
\arguments{
\item{destfile}{A character (default: "/tmp/tmp.hdf5"). A file name to download.}

\item{event_name}{A character. A specific event name.}

\item{dur}{An integer (default: 32). A duration in second of the data.}

\item{file.format}{A character (default: "hdf5"). A file format. (one of "gwf", "hdf5", or "txt")}

\item{sampling.freq}{A numeric (default: 4096). A sampling frequency of the data.}

\item{timeout}{An integer (default: 300). To extend timeout for downloading.}

\item{load}{A logical (default: FALSE). Whether it assigns the data directly into the R environment. 'file.format' must be "hdf5".}

\item{remove}{A logical (default: FALSE). Whether it removes data file after loading.}

\item{direct.url}{A character (default: NULL). The file URL can be directly given.}

\item{detector}{A character. A specific detector name.}
}
\value{
A list. $Data: A vector of strain, $tstart: A numeric of GPS start time, if 'load=TRUE'.
}
\description{
Download strain data from GWOSC
}
