% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{DQ_Dec2Bits}
\alias{DQ_Dec2Bits}
\title{Convert Decimal DQ Flag to Bit Representation}
\usage{
DQ_Dec2Bits(decimal.dq, len = 7)
}
\arguments{
\item{decimal.dq}{An integer. Decimal representation of the DQ flag.}

\item{len}{An integer (default: 7). Number of bits to extract.}
}
\value{
An integer vector of length \code{len} containing bitwise flags.
}
\description{
Convert Decimal DQ Flag to Bit Representation
}
