% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{DQ_ShortNames}
\alias{DQ_ShortNames}
\title{List of DQ Bit Names with Their Bit Positions}
\usage{
DQ_ShortNames()
}
\value{
An integer-named character vector indicating short names and their bit index.
}
\description{
List of DQ Bit Names with Their Bit Positions
}
\details{
The default mapping is:
\itemize{
  \item \code{DATA}       = 0
  \item \code{CBC_CAT1}   = 1
  \item \code{CBC_CAT2}   = 2
  \item \code{CBC_CAT3}   = 3
  \item \code{BURST_CAT1} = 4
  \item \code{BURST_CAT2} = 5
  \item \code{BURST_CAT3} = 6
}
}
