% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{DQlev}
\alias{DQlev}
\title{Extract Specific DQ Level from Decimal Flag}
\usage{
DQlev(dq, level = "BURST_CAT2")
}
\arguments{
\item{dq}{An integer. Decimal DQ flag.}

\item{level}{A character string. DQ level to extract (e.g., \code{"BURST_CAT2"} or \code{"all"}).}
}
\value{
If \code{level = "all"}, returns a named integer vector of bit flags.\cr
Otherwise, returns a scalar (0 or 1) indicating the presence of the selected DQ level.
}
\description{
Extract Specific DQ Level from Decimal Flag
}
