% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{dur}
\alias{dur}
\title{Get Duration of the Original Time Series}
\usage{
dur(x)
}
\arguments{
\item{x}{An `fs` object derived from a `ts` object.}
}
\value{
A numeric scalar. Duration in seconds.
}
\description{
Computes the total duration in seconds of the time-domain signal that
generated the `fs` object.
}
