% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{ear}
\alias{ear}
\title{Ensemble AR model fitting across multiple orders}
\usage{
ear(
  ts,
  ps,
  aic = T,
  return.vec = T,
  return.var = T,
  return.mean = T,
  collector = "median",
  ...
)
}
\arguments{
\item{ts}{A `ts` object.}

\item{ps}{Vector of AR orders.}

\item{aic}{Use AIC selection.}

\item{return.vec}{Flatten features.}

\item{return.var}{Include prediction variance.}

\item{return.mean}{Include series mean.}

\item{collector}{Aggregation method.}

\item{...}{Additional args passed to `burgar()`.}
}
\value{
Residual `ts` object with ensemble features.
}
\description{
Ensemble AR model fitting across multiple orders
}
