% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{embedParallelCpp}
\alias{embedParallelCpp}
\title{C++ version of embed function}
\usage{
embedParallelCpp(x, dimension, numCores)
}
\arguments{
\item{x}{A numeric vector.}

\item{dimension}{A numeric.}

\item{numCores}{A numeric. A number of cores}
}
\description{
C++ version of embed function
}
