% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{eoa}
\alias{eoa}
\title{Ensemble of Averages (EoA)}
\usage{
eoa(ts, qs, collector = "median")
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{qs}{A vector of length 2 whose first value is lower bound of q and second is upper bound of q for 'ma'.}

\item{collector}{A character (default: "median"). A function name for collecting various MA models. (one of "mean", "median", or "pca)}
}
\value{
A list. \code{$res}: A EoA smoothed time series (`ts`) object. `$df`: A data frame containing various MA(q) along the column.
}
\description{
Ensemble of Averages (EoA)
}
