% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{evenify}
\alias{evenify}
\title{If length(Data$data) is odd, cut the last element}
\usage{
evenify(ts)
}
\arguments{
\item{ts}{A time series (`ts`) object.}
}
\value{
A time series (`ts`) object with even length.
}
\description{
If length(Data$data) is odd, cut the last element
}
