% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{extract_pc}
\alias{extract_pc}
\title{Extract principal component(s) from input}
\usage{
extract_pc(x, pc = "PC1")
}
\arguments{
\item{x}{A data frame or matrix.}

\item{pc}{A string or vector of PC names (e.g., "PC1").}
}
\value{
A matrix or vector of selected components.
}
\description{
Extract principal component(s) from input
}
