% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{floor.digit}
\alias{floor.digit}
\title{Floor with Decimal Digits}
\usage{
\method{floor}{digit}(x, digits = 0)
}
\arguments{
\item{x}{A numeric value or vector.}

\item{digits}{Integer (default: 0). Number of decimal digits to retain.}
}
\value{
A numeric vector with values floored to specified digits.
}
\description{
Applies the \code{floor()} function to a number with precision controlled by number of digits.
}
