% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{freqs}
\alias{freqs}
\title{Get Frequency Samples from `fs`}
\usage{
freqs(x)
}
\arguments{
\item{x}{An `fs` object.}
}
\value{
A numeric vector of frequency values (in Hz).
}
\description{
Computes the discrete frequency values corresponding to the bins of an `fs` object.
}
