% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{fs}
\alias{fs}
\title{fs: FrequencySeries class}
\usage{
fs(x, df)
}
\arguments{
\item{x}{A numeric vector.}

\item{df}{A numeric. A delta frequency.}
}
\description{
fs: FrequencySeries class
}
\examples{
# fs(c(1,2,3), df=0.5)

}
