% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{fs_df}
\alias{fs_df}
\title{Convert `fs` to Data Frame}
\usage{
fs_df(x)
}
\arguments{
\item{x}{An `fs` object.}
}
\value{
A data frame with two columns: `freqs` and `PSD`.
}
\description{
Converts an `fs` object to a data frame with columns `freqs` and `PSD`.
}
