% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoC.R
\name{gen.data}
\alias{gen.data}
\title{Generating random noise and inserting signal into the noise}
\usage{
gen.data(
  signal,
  sampling_freq,
  beta = 1,
  tlen = 32,
  pre.sec = 2,
  fl = 5,
  delta_f = 1/32,
  noise_type = "aLIGOZeroDetHighPower",
  txt_path = NULL,
  t0.signal = F,
  noise.seed = NULL
)
}
\description{
Generating random noise and inserting signal into the noise
}
