% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoC.R
\name{gen_noise}
\alias{gen_noise}
\title{Generate noise sample}
\usage{
gen_noise(
  tlen,
  sampling.freq,
  fl = 5,
  delta_f = 1/16,
  noise_type = "aLIGOZeroDetHighPower",
  txt_path = NULL,
  noise.mean = 0,
  noise.sd = 1,
  seed = NULL
)
}
\description{
Generate noise sample
}
