% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detector.R
\name{get.available.detectors}
\alias{get.available.detectors}
\title{Get available GW detectors}
\usage{
get.available.detectors()
}
\value{
A data frame. Output includes columns of "Detector name", "Abbreviation" of detector name, latitude and longitude of the detector.
}
\description{
Get available GW detectors
}
