% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pycbc.R
\name{get_available_detectors}
\alias{get_available_detectors}
\title{Get available GW detectors}
\usage{
get_available_detectors()
}
\value{
A data frame. Output includes columns of "Detector name", "Abbreviation" of detector name, latitude and longitude of the detector.
}
\description{
Get available GW detectors
}
