% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{get_break_rngdgt}
\alias{get_break_rngdgt}
\title{Compute break points with rounded range}
\usage{
get_break_rngdgt(x, bin = 0.2)
}
\arguments{
\item{x}{A numeric vector of values to compute breaks for.}

\item{bin}{A numeric scalar indicating the bin width (default: 0.2).}
}
\value{
A numeric vector of break points from minimum to maximum of \code{x}, rounded to fit the bin width.
}
\description{
Compute break points with rounded range
}
