% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{get_cutoff_indices}
\alias{get_cutoff_indices}
\title{Get two boundary indices for frequency range cropping}
\usage{
get_cutoff_indices(flow, fupp, df, N)
}
\description{
Reference: https://pycbc.org/pycbc/latest/html/_modules/pycbc/filter/matchedfilter.html#get_cutoff_indices
}
