% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{get_gps}
\alias{get_gps}
\title{Add GPS time column to a data frame}
\usage{
get_gps(df, ref.ts)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble}. The data to which the GPS time will be added.}

\item{ref.ts}{A \code{ts} object. Reference time series from which GPS timestamps are extracted.}
}
\value{
A \code{tibble} identical to \code{df} but with an additional \code{GPS} column inserted after the \code{time} column.
}
\description{
Adds a \code{GPS} column to a data frame by extracting timestamps from a reference \code{ts} object.
}
