% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSC.R
\name{get_gwosc}
\alias{get_gwosc}
\title{Get information from GWOSC!}
\usage{
get_gwosc(offline = F, csvpath = NULL)
}
\arguments{
\item{offline}{A logical (default: FALSE). Whether it searchs online or gets information from pre-downloaded csv file offline.}

\item{csvpath}{A character. A pre-downloaded csv file path. If 'offline=TRUE', this need to be given.}

\item{event_name}{A character (default: "ALL).}
}
\value{
A data frame.
}
\description{
Get information from GWOSC!
}
