% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{get_last_tcen}
\alias{get_last_tcen}
\title{Extract the most recent cluster center time}
\usage{
get_last_tcen(proc)
}
\arguments{
\item{proc}{A dataframe output from the anomaly detection pipeline, containing
at least the columns `cluster` and `t_cen`.}
}
\value{
A numeric value representing the last cluster's center time (`t_cen`).
}
\description{
Retrieves the central time (`t_cen`) of the last (i.e., most recent) cluster
from a processed anomaly dataframe. If no cluster exists, it attempts to retrieve
`prev_tcen` from the parent environment.
}
\note{
Assumes that `prev_tcen` is defined in the parent frame if no clusters are found.
}
