% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{get_limit}
\alias{get_limit}
\title{Get limit range from ts by 110 % of maximum value}
\usage{
get_limit(x, mar.frac = 1.5)
}
\arguments{
\item{x}{A numeric vector.}

\item{mar.frac}{A numeric. A margin. This will be multiplied on amax(x).}
}
\value{
A vector. c(-amax, +amax)*mar.frac.
}
\description{
Get limit range from ts by 110 % of maximum value
}
