% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{get_MissingValues}
\alias{get_MissingValues}
\title{Get Missing Values (NA) between ts and ref in terms of time stamps}
\usage{
get_MissingValues(ts, ref)
}
\arguments{
\item{ts}{A time series (`ts`) object who has NA values.}

\item{ref}{A time series (`ts`) object for comparison.}
}
\value{
A list with
     \code{$all}: time stamps whose values are NA,
     \code{$head}: time stamps whose values are NA at the head,
     \code{$tail}: time stamps whose values are NA at the tail.
}
\description{
Get Missing Values (NA) between ts and ref in terms of time stamps
}
