% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{get_order}
\alias{get_order}
\title{Calculate the order of magnitude}
\usage{
get_order(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric. The order of magnitude of the maximum absolute value.
}
\description{
Calculate the order of magnitude
}
