% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSC.R
\name{get_timeline}
\alias{get_timeline}
\title{Retrieve timeline segments for a run/timeline (may include duty_cycle)}
\usage{
get_timeline(
  obsrun,
  timeline,
  GPSstart,
  GPSend,
  compact = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{obsrun}{Character run id (e.g. "O1").}

\item{timeline}{Character timeline id (e.g. "H1_DATA" or "H1_BURST_CAT2").}

\item{GPSstart}{Integer start GPS.}

\item{GPSend}{Integer stop GPS.}

\item{compact}{Logical If TRUE request compact form when supported.}

\item{verbose}{Logical for debugging.}
}
\value{
data.frame with columns start, stop, and duty_cycle (if present).
}
\description{
Query GWOSC timeline segments:
/api/v2/runs/<obsrun>/timelines/<timeline>/segments
}
\examples{
\dontrun{
segs <- get_timeline("O1", "H1_DATA", 1126051217, 1126051217 + 86400)
}
}
