% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{gps2utc}
\alias{gps2utc}
\title{Convert GPS time to UTC}
\usage{
gps2utc(
  x,
  origin = as.Date("1980-01-06"),
  out_format = "\%Y-\%m-\%d \%H:\%M:\%OS"
)
}
\arguments{
\item{x}{A numeric. GPS time.}

\item{origin}{A character or Date (default: "1980-01-06"). GPS epoch.}

\item{out_format}{A character. Output format string.}
}
\value{
A character string of UTC time.
}
\description{
Convert GPS time to UTC
}
