% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{init_pipe}
\alias{init_pipe}
\title{Initialize pipeline environment}
\usage{
init_pipe(dets = c("H1", "L1"))
}
\arguments{
\item{dets}{A character vector of detector names. Defaults to \code{c("H1", "L1")}.}
}
\value{
None. This function assigns \code{prev_batch}, \code{res.net}, and \code{coinc.lis}
into the calling environment.
}
\description{
Initializes global pipeline variables required for the anomaly detection pipeline,
including previous batch placeholders, detector-wise result containers, and coincidence list.
This function is designed to be run at the beginning of a `pipe()` execution.
}
