% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{inner}
\alias{inner}
\title{Compute inner product between two vectors}
\usage{
inner(x, y)
}
\arguments{
\item{x}{A complex or numeric vector.}

\item{y}{A complex or numeric vector.}
}
\value{
Complex scalar. Inner product of `x` and `y`.
}
\description{
Compute inner product between two vectors
}
