% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QT.R
\name{interp2d}
\alias{interp2d}
\title{Surface Interpolation}
\usage{
interp2d(x, y, z, xout, yout, method = "linear")
}
\arguments{
\item{x}{A numeric vector to be interpolated.}

\item{y}{A numeric vector to be interpolated.}

\item{z}{A numeric matrix to be interpolated. `dim(z) == c(length(x), length(y))`.}

\item{xout}{A numeric vector. The output grid.}

\item{yout}{A numeric vector. The output grid.}

\item{method}{A character. Available methods is `linear` or `spline`.}
}
\value{
`zout`, the interpolated 2d matrix.
}
\description{
A mutated function of `fields::interp.surface` and `fields::interp.surface.grid`. This function is boosted up by `sapply`.
}
