% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{is.anomdet}
\alias{is.anomdet}
\title{Check if anomalies were detected}
\usage{
is.anomdet(proc)
}
\arguments{
\item{proc}{A data frame that includes an `anomaly` column with binary flags (0 or 1).}
}
\value{
A logical value: `TRUE` if any anomalies are detected, `FALSE` otherwise.
}
\description{
Returns `TRUE` if the given `proc` data frame contains any rows with `anomaly == 1`.
}
