% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{list.append}
\alias{list.append}
\title{Append or assign a value into a nested list}
\usage{
list.append(list, where, value, name = NULL)
}
\arguments{
\item{list}{A list object with a sublist at \code{where}.}

\item{where}{A character string indicating the name of the sublist to modify.}

\item{value}{The value to be appended or assigned.}

\item{name}{Optional. If provided, the value is assigned under this name. Otherwise, it is appended as unnamed.}
}
\value{
The modified list object.
}
\description{
Adds a new element to a nested list, either as an unnamed entry appended to the end
or as a named entry assigned under a specified key.
}
