% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSC.R
\name{list_gwosc_param}
\alias{list_gwosc_param}
\title{List all available GWOSC parameter names}
\usage{
list_gwosc_param()
}
\value{
A character vector of parameter names.
}
\description{
Returns a character vector of all valid parameter names
that can be used in \code{\link{get_gwosc_param}}.
}
\examples{
list_gwosc_param()
}
\seealso{
\code{\link{get_gwosc_param}}
}
