% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSC.R
\name{list_timelines}
\alias{list_timelines}
\title{List timelines available for a given run}
\usage{
list_timelines(obsrun, verbose = FALSE)
}
\arguments{
\item{obsrun}{character run id, e.g. "O1"}

\item{verbose}{logical for debugging}
}
\value{
character vector of timeline names
}
\description{
Return timelines (e.g. "H1_DATA", "H1_BURST_CAT2") available for a run.
}
