% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{ma}
\alias{ma}
\title{Two-sided Moving Average (MA) smoother}
\usage{
ma(ts, order, na.rm = T)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{order}{An integer for the order of MA.}

\item{na.rm}{A logical whether remove NA values or not.}
}
\value{
A MA smoothed time series (`ts`) object.
}
\description{
Two-sided Moving Average (MA) smoother
}
