% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{mae}
\alias{mae}
\title{Median Absolute Error for sample}
\usage{
mae(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric. A sampled median absolute error.
}
\description{
Median Absolute Error for sample
}
