% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{matched.filter}
\alias{matched.filter}
\title{Matched filter (R ver.)}
\usage{
matched.filter(template, data, psd = NULL, fl = NULL, fu = NULL, h.norm = NULL)
}
\arguments{
\item{template}{A `ts` object of waveform template.}

\item{data}{A `ts` object of data.}

\item{psd}{A `fs` object of PSD.}

\item{fl}{A numeric. Lower frequency cutoff.}

\item{fu}{A numeric. Upper frequency cutoff.}

\item{h.norm}{A custom normalization factor.}
}
\value{
A list containing: `snr.ts`, `snr.fs`, and `norm`, where SNR in `ts` and `fs` object and the normalization factor, respectively. Final SNR will be \code{norm*abs(snr.ts)}
}
\description{
I've digged out the source codes to translate from Python to R.
}
