% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{message_by}
\alias{message_by}
\title{Message out by each iteration or No verbose}
\usage{
message_by(i = i, verbose = T, msg)
}
\arguments{
\item{i}{An expression. A iterable arguments in for loop.}

\item{verbose}{A logical or An integer. If 'verbose=TRUE', it message out by every iteration. Else 'verbose=integer()', it message out by every verbose-th iteration.}

\item{msg}{A character. A message to print out.}
}
\description{
Message out by each iteration or No verbose
}
