% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{message_verb}
\alias{message_verb}
\title{Conditionally Print Message}
\usage{
message_verb(..., v = TRUE)
}
\arguments{
\item{...}{Character strings to print.}

\item{v}{Logical; whether to print the message.}
}
\value{
None.
}
\description{
Print message only when \code{v = TRUE}.
}
