% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{mmed}
\alias{mmed}
\title{Simple Median filter wrapper}
\usage{
mmed(ts, order)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{order}{An integer. The order of median filter.}
}
\value{
A median filtered time series (`ts`) object.
}
\description{
Simple Median filter wrapper
}
