% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{MovingAverage}
\alias{MovingAverage}
\title{Apply MA or EoA depending on `q`}
\usage{
MovingAverage(x, q, verbose = TRUE, ...)
}
\arguments{
\item{x}{A `ts` object.}

\item{q}{MA order(s).}

\item{verbose}{Verbose output.}

\item{...}{Additional args passed to `eoa()`.}
}
\value{
Smoothed `ts` object.
}
\description{
Apply MA or EoA depending on `q`
}
