% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{one.ts}
\alias{one.ts}
\title{Scale ts}
\usage{
one.ts(ts)
}
\arguments{
\item{ts}{A time series (`ts`) object.}
}
\value{
A scaled time series. A scale factor is calculated by '`get.order`'. The factor will be stored in '`$order`' attribute.
}
\description{
Scale ts
}
