% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{one_ts}
\alias{one_ts}
\title{Normalize a `ts` to unit scale}
\usage{
one_ts(ts)
}
\arguments{
\item{ts}{A `ts` object.}
}
\value{
A normalized `ts` object. Scale factor is stored in \code{attr(..., "order")}.
}
\description{
Normalize a `ts` to unit scale
}
