% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{oscillo.option}
\alias{oscillo.option}
\title{Common ggplot options for oscillogram}
\usage{
oscillo.option(ts, tzero = 0, title = NULL)
}
\arguments{
\item{ts}{A time series (ts) object.}

\item{tzero}{A numeric (default: 0). A zero time. Times will be shifted by 'tzero'.}

\item{title}{A character. A title for the plot.}
}
\value{
A list. This can be added to ggplot object to apply same theme.
}
\description{
For applying common theme.
}
