% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{overlap}
\alias{overlap}
\title{Calculate real overlap}
\usage{
overlap(
  vec1,
  vec2,
  psd = NULL,
  low_frequency_cutoff = NULL,
  high_frequency_cutoff = NULL,
  normalized = TRUE
)
}
\arguments{
\item{vec1}{A `ts` or `fs`. The input vector containing a waveform.}

\item{vec2}{A `ts` or `fs`. The input vector containing a waveform.}

\item{psd}{A `fs` object (default: `NULL`). A power spectral density to weight the overlap.}

\item{low_frequency_cutoff}{A numeric (default: `NULL`). The frequency to begin the overlap.}

\item{high_frequency_cutoff}{A numeric (default: `NULL`). The frequency to stop the overlap.}

\item{normalized}{A logical (default: `TRUE`). Set if the overlap is normalized. If true, it will range from 0 to 1.}
}
\value{
overlap: numeric
}
\description{
Return the overlap between the two TimeSeries or FrequencySeries
}
