% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{plot.fs}
\alias{plot.fs}
\title{Plot function for fs class}
\usage{
\method{plot}{fs}(
  fs,
  log = "xy",
  xlab = "Frequency (Hz)",
  ylab = "PSD",
  xlim = NULL,
  ylim = NULL
)
}
\description{
Plot function for fs class
}
