% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{plot_aic}
\alias{plot_aic}
\title{Plot AIC curve for AR model}
\usage{
plot_aic(ar)
}
\arguments{
\item{ar}{An `ar` object.}
}
\value{
A ggplot AIC plot with zoomed-in inset.
}
\description{
Plot AIC curve for AR model
}
