% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{plot_anomalies_multi}
\alias{plot_anomalies_multi}
\title{Plot anomalies for multiple detectors}
\usage{
plot_anomalies_multi(anom.det, ...)
}
\arguments{
\item{anom.det}{A data.frame with anomaly results that includes detector info in column "det".}

\item{...}{Passed to \code{plot_anomalies()}.}
}
\value{
A faceted ggplot object showing anomaly results per detector.
}
\description{
Plot anomalies for multiple detectors
}
