% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{plot_background}
\alias{plot_background}
\title{Plot histogram of background detection statistic}
\usage{
plot_background(
  S.bkg.vec,
  color = "black",
  binwidth = 0.2,
  xlimit = c(7, 30),
  ylimit = c(0.005, 500),
  factor = 1
)
}
\arguments{
\item{S.bkg.vec}{A numeric vector of background detection statistic (e.g., \eqn{\mathcal{S}} values).}

\item{color}{A character string specifying histogram color.}

\item{binwidth}{Numeric. Width of histogram bins (default: 0.2).}

\item{xlimit}{Numeric vector of length 2. Limits of x-axis (default: c(7, 30)).}

\item{ylimit}{Numeric vector of length 2. Limits of y-axis (log10 scale, default: c(5e-3, 5e2)).}

\item{factor}{Numeric. Rescaling factor for counts (default: 1).}
}
\value{
A list of one ggplot object named \code{histogram}.
}
\description{
Plot histogram of background detection statistic
}
