% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{plot_coinc}
\alias{plot_coinc}
\title{Plot coincidence significance values over time}
\usage{
plot_coinc(
  coinc.res,
  tzero = NULL,
  p_crit = 0.05,
  a = 3,
  legend.position = "tr",
  annotate.vals = FALSE,
  annotate.thresh = p_crit
)
}
\arguments{
\item{coinc.res}{A data.frame returned from \code{coincide_P0()}.}

\item{tzero}{Optional GPS time to align x-axis (default: first timestamp in \code{coinc.res}).}

\item{p_crit}{Critical p-value to draw significance threshold (default: 0.05).}

\item{a}{Scaling factor for significance function (default: 3).}

\item{legend.position}{Position of legend inside plot. One of "tr", "tl", "br", "bl".}

\item{annotate.vals}{A logical. Annotate values above threshold (default: FALSE).}
}
\value{
A ggplot object with time-series of coincidence significance.
}
\description{
Plot coincidence significance values over time
}
