% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{plot_oscillo}
\alias{plot_oscillo}
\title{Plot oscillogram}
\usage{
plot_oscillo(
  ts,
  tzero = NULL,
  trange = NULL,
  ylim = "pm",
  title = NULL,
  size = 0.3,
  ...
)
}
\arguments{
\item{ts}{A time series (ts) object.}

\item{tzero}{A numeric (default: 0). A zero time. Times will be shifted by 'tzero'.}

\item{trange}{A vector. A time range for x axis limit.}

\item{ylim}{A character of "pm" (default) or A vector. "pm": from \code{-1.1*max(abs(ts))} to \code{1.1*max(abs(ts))}. \code{c(lwr, upr)}: from lwr to upr.}

\item{title}{A character. A title for the plot.}

\item{...}{Additional arguments for '\code{ggplot2::geom_line}'.}

\item{color}{A character. A line color.}

\item{width}{A numeric. A line width.}
}
\value{
A ggplot object.
}
\description{
Plot oscillogram
}
