% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{plot_oscillo_multi}
\alias{plot_oscillo_multi}
\title{Plot multi-panel oscillograms by facet_wrap}
\usage{
plot_oscillo_multi(
  ts.df,
  tzero = NULL,
  trange = NULL,
  facet.scale = "fixed",
  facet.label = NULL,
  order.of.mag = TRUE,
  title = NULL,
  color = NULL,
  width = 0.3,
  ...
)
}
\arguments{
\item{ts.df}{A data frame. This should be generated by 'ts.df' or contains 'time' column.}

\item{tzero}{A numeric (default: 0). A zero time. Times will be shifted by 'tzero'.}

\item{trange}{A vector. A time range for x axis limit.}

\item{title}{A character. A title for the plot.}

\item{color}{A character. A line color.}

\item{width}{A numeric. A line width.}

\item{...}{Additional arguments for `ggplot2::geom_line`.}

\item{facet_scale}{A character. How to scale each facets. (One of "fixed", "free_y", or "free_x")}

\item{ylim}{A character of "pm" (default) or A vector. "pm": from `-1.1*max(abs(ts))` to `1.1*max(abs(ts))`. `c(lwr, upr)`: from lwr to upr.}
}
\value{
A ggplot object.
}
\description{
Plot multi-panel oscillograms by facet_wrap
}
