% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{print.fs}
\alias{print.fs}
\title{Custom Print Method for `fs` Class}
\usage{
\method{print}{fs}(x, ..., ts.info = TRUE)
}
\arguments{
\item{x}{An `fs` object.}

\item{...}{Additional arguments (ignored).}

\item{ts.info}{Logical. Whether to show associated time-series info (default: TRUE).}
}
\value{
None. Used for side effects.
}
\description{
Prints summary information for an `fs` object, including delta_f, flen,
and associated time-domain metadata if available.
}
