% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detector.R
\name{psd.noise}
\alias{psd.noise}
\title{Gaussian noise from detetctor PSD}
\usage{
psd.noise(
  det,
  duration,
  tstart,
  sampling.freq = 4096,
  fl = 15,
  delta_f = 1/32,
  seed = NULL
)
}
\arguments{
\item{det}{A character. An abbreviation of detector name.
(this ver., H1, L1, V1, K1, or E1)}

\item{duration}{A numeric. A length of noise in second.}

\item{tstart}{A numeric. Starting time of the noise.}

\item{sampling.freq}{A numeric (default: 4096). A sampling rate in Hz.}

\item{fl}{A numeric (default: 32). A lower frequency cutoff.}

\item{delta_f}{A numeric (default: 1/32). A frequency resolution.}

\item{seed}{A numeric (default: NULL). Random seed.}
}
\value{
ts object of Gaussian noise.
}
\description{
Gaussian noise from detetctor PSD
}
