% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pycbc.R
\name{psd_noise}
\alias{psd_noise}
\title{Simulate Gaussian noise from detector PSD}
\usage{
psd_noise(
  det,
  duration,
  tstart,
  sampling.freq = 4096,
  fl = 15,
  delta_f = 1/32,
  seed = NULL
)
}
\arguments{
\item{det}{A character. Detector abbreviation ("H1", "L1", "V1", "K1", or "E1").}

\item{duration}{A numeric. Noise duration (in seconds).}

\item{tstart}{A numeric. Start GPS time of the noise.}

\item{sampling.freq}{A numeric. Sampling frequency (Hz). Default: 4096.}

\item{fl}{A numeric. Lower frequency cutoff. Default: 15 Hz.}

\item{delta_f}{A numeric. Frequency resolution. Default: 1/32.}

\item{seed}{Optional numeric. Random seed for reproducibility.}
}
\value{
A \code{ts} object of simulated Gaussian noise.
}
\description{
Generate colored Gaussian noise with power spectral density of a given detector.
}
\details{
This function interfaces with Python's \pkg{pycbc} library using \pkg{reticulate}.
Please ensure that the required Python packages are available in your environment.
}
