% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTrasnform.R
\name{qplane}
\alias{qplane}
\title{Q-plane}
\usage{
qplane(qplane.tile.list, fseries, return_complex = FALSE)
}
\arguments{
\item{qplane.tile.list}{A list containing a vector of q-tile for each q-plane.}

\item{fseries}{A `fs` object. frequency-series data set.}

\item{return_complex}{A logical. Return the raw complex series instead of the normalized power.}
}
\value{
A list containing q: The q of the maximum q plane, times: The time that the qtransform is sampled, freqs: The frequencies that the qtransform is sampled, and qplane (2d): The two dimensional interpolated qtransform of this time series.
}
\description{
Q-plane
}
