% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QT.R
\name{qseries}
\alias{qseries}
\title{Perform Q-transform on a single tile}
\usage{
qseries(fseries, Q, f0, return_complex = FALSE)
}
\arguments{
\item{fseries}{An \code{fs} object. Frequency-series representation of the data.}

\item{Q}{A numeric. Q value of the tile.}

\item{f0}{A numeric. Center frequency of the tile.}

\item{return_complex}{Logical (default: FALSE). If \code{TRUE}, return complex time series instead of normalized energy.}
}
\value{
A \code{ts} object representing either normalized energy or complex waveform.
}
\description{
Compute Q-transform for a single Q and center frequency tile.
}
