% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTrasnform.R
\name{qtransform}
\alias{qtransform}
\title{Q-transform (R ver.)}
\usage{
qtransform(
  ts,
  delta_t = NULL,
  delta_f = NULL,
  logfsteps = NULL,
  frange = NULL,
  qrange = c(4, 64),
  mismatch = 0.2,
  return_complex = FALSE
)
}
\arguments{
\item{ts}{A `ts` object.}

\item{delta_t}{A numeric. The time resolution to interpolate to.}

\item{delta_f}{A numeric. The frequency resolution to interpolate to.}

\item{logfsteps}{A numeric. Do a log interpolation (incompatible with delta_f option) and set the number of steps to take.}

\item{frange}{A numeric vector. frequency range.}

\item{qrange}{A numeric vector. q range.}

\item{mismatch}{A numeric (default: 0.2). Mismatch between frequency tiles.}

\item{return_complex}{A logical (default: FALSE). return the raw complex series instead of the normalized power.}
}
\value{
A list containing times: The time that the qtransform is sampled, freqs: The frequencies that the qtransform is sampled, and qplane (2d): The two dimensional interpolated qtransform of this time series.
}
\description{
Q-transform (R ver.)
}
