% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{range.width}
\alias{range.width}
\title{Rounded Range with Specified Bin Width}
\usage{
\method{range}{width}(x, width)
}
\arguments{
\item{x}{A numeric vector.}

\item{width}{Numeric. Desired bin width.}
}
\value{
A numeric vector of length 2 indicating start and end of adjusted range.
}
\description{
Computes a range that aligns with a given bin width, rounding outwards to ensure coverage.
}
