% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{read.H5}
\alias{read.H5}
\title{Read H5 file}
\usage{
read.H5(file, sampling.freq, dq.level = "BURST_CAT2")
}
\arguments{
\item{file}{A character. A file path to read.}

\item{sampling.freq}{A numeric. A sampling frequency.}
}
\value{
A ts object.
}
\description{
Read H5 file
}
