% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{read_DQ}
\alias{read_DQ}
\title{Read Only the DQ Mask from HDF5 File}
\usage{
read_DQ(file, dq.level = "BURST_CAT2")
}
\arguments{
\item{file}{A character string. Path to the HDF5 file.}

\item{dq.level}{A character string. DQ level to extract (default: "BURST_CAT2").
Use \code{"all"} to extract all DQ bits.}
}
\value{
A \code{ts} object representing the DQ mask.
}
\description{
Read Only the DQ Mask from HDF5 File
}
