% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detector.R
\name{relpass.time}
\alias{relpass.time}
\title{Time source gravitational-wave passes through detector}
\usage{
relpass.time(deltat, dets, ra, dec, t_gps, verbose = T)
}
\arguments{
\item{dets}{A vector. Characters of detector abbreviations.}

\item{ra}{A numeric. RA of the GW source in radian.}

\item{dec}{A numeric. Dec of the GW source in radian.}

\item{t_gps}{A numeric. Passing time of the GW source to Earth in gps second.}
}
\value{
A numeric. (If 'dets' has one argument) 
        A data frame. (If 'dets' is multiple)
}
\description{
Time source gravitational-wave passes through detector
}
