% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pycbc.R
\name{relpass_time}
\alias{relpass_time}
\title{Relative arrival times of signal at detectors}
\usage{
relpass_time(deltat, dets, ra, dec, t_gps, verbose = T)
}
\arguments{
\item{deltat}{A numeric. Time resolution to round delays (in seconds).}

\item{dets}{A character vector of detector abbreviations.}

\item{ra}{A numeric. Right Ascension of the source (in radians).}

\item{dec}{A numeric. Declination of the source (in radians).}

\item{t_gps}{A numeric. GPS time of signal arrival at Earth's center.}

\item{verbose}{Logical. If \code{TRUE}, print intermediate results.}
}
\value{
If single detector, a list with \code{time_rel} and \code{remainder}. If multiple detectors, a list with matrix \code{dt_rel}, absolute delays \code{dt_earth}, and \code{remainder}.
}
\description{
Compute arrival time delays among detectors for a given sky position.
}
\details{
This function interfaces with Python's \pkg{pycbc} library using \pkg{reticulate}.
Please ensure that the required Python packages are available in your environment.
}
