% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{residual}
\alias{residual}
\title{Compute residuals using AR coefficients}
\usage{
residual(x, ar, numCores = parallel::detectCores())
}
\arguments{
\item{x}{Time series data.}

\item{ar}{AR coefficients.}

\item{numCores}{Number of CPU cores to use.}
}
\value{
Residual time series.
}
\description{
Compute residuals using AR coefficients
}
