% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{resize}
\alias{resize}
\title{Resize a `ts` to target length}
\usage{
resize(x, nlen, align = c("left", "center", "right"))
}
\arguments{
\item{x}{A `ts` object.}

\item{nlen}{An integer. Desired length.}

\item{align}{Character. One of "left", "center", "right".
Determines how the series is aligned when padding/truncating.
Default: "left".}
}
\value{
A resized `ts` object (padded with zeros or truncated).
}
\description{
Resize a `ts` to target length
}
