% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{sar}
\alias{sar}
\title{Simple wrapper for AR fitting}
\usage{
sar(ts, ...)
}
\arguments{
\item{ts}{A `ts` object.}

\item{...}{Arguments passed to `burgar()`.}
}
\value{
Residual `ts` object with metadata.
}
\description{
Simple wrapper for AR fitting
}
