% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{savefig}
\alias{savefig}
\title{Save ggplot plot}
\usage{
savefig(plot, file, width = 8, height = 5, show = T, verbose = F, ...)
}
\arguments{
\item{plot}{A ggplot object.}

\item{file}{A character. A file path for save.}

\item{width}{A numeric (default: 8). Width of plot size.}

\item{show}{A logical (default: TRUE). Whether it shows and returns the plot on pane after save.}

\item{...}{Additional arguments for '\code{ggpubr::ggexport}'.}

\item{heigth}{A numeric (default: 5). Height of plot size.}
}
\value{
A ggplot object, if \code{show=TRUE}.
}
\description{
Save ggplot plot
}
