% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{savemsg}
\alias{savemsg}
\title{Save message to txt file}
\usage{
savemsg(msg, ..., file, show = T)
}
\arguments{
\item{msg}{A character. A message to save.}

\item{...}{A character. Additional messages to save.}

\item{file}{A character. A file path for save.}

\item{show}{A logical (default: TRUE). Whether it prints the messages on console.}
}
\description{
Save message to txt file
}
