% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{savetab}
\alias{savetab}
\title{Save table as txt file}
\usage{
savetab(
  x,
  file,
  signif = 3,
  col.names = F,
  row.names = F,
  quote = F,
  show = T,
  append = F,
  nsep = 30,
  ...
)
}
\arguments{
\item{x}{A data frame.}

\item{file}{A character. A file path for save.}

\item{signif}{An integer (default: 3). A significant digits.}

\item{col.names}{A logical (default: FALSE). Whether it saves column names together.}

\item{quote}{A logical (default: FALSE). Whether it saves each elements with quote.}

\item{show}{A logical (default: TRUE). Whether it prints the table on console.}

\item{append}{A logical (default: FALSE). Whether it appends on an existing file.}

\item{nsep}{An integer (default: 30). If '\code{append=TRUE}', it adds separation dashed line with the length of 'nsep' before adding.}

\item{...}{Additional arguments for '\code{insight::export_table}'.}

\item{row.naems}{A logical (default: FALSE). Whether it saves row names together.}
}
\description{
This uses 'insight' package to convert a data frame to neat "Markdown" version of table.
}
