% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{seqarima}
\alias{seqarima}
\title{Sequential ARIMA function}
\usage{
seqarima(
  ts,
  d = NULL,
  p = NULL,
  q = NULL,
  t.seg = 0.5,
  d_max = NULL,
  p_max = NULL,
  min.q = 1,
  q_max = NULL,
  fl = NULL,
  fu = NULL,
  ar.choose = "residual",
  var.method = 1,
  numCores = NULL,
  ma.collector = "median",
  return.mas = FALSE,
  elap.t = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{d}{An integer. The FIXED order of difference.}

\item{p}{An integer. The FIXED order of AR.}

\item{q}{An integer. The FIXED order of MA.}

\item{t.seg}{A numeric (default: 0.5). A time length in each segment.}

\item{d_max}{An integer (default: NULL). If d_max is given, the order d will not exceed d_max.}

\item{p_max}{An integer. The maximum order of AR.}

\item{min.q}{An integer (default: 1). The lower bound of q for MAC.}

\item{q_max}{An integer. The upper bound of q for MAC.}

\item{fl}{A numeric. The frequency lower bound for the band-pass filter.}

\item{fu}{A numeric. The frequency upper bound for the band-pass filter.}

\item{ar.choose}{A character (default: "residual"). Choose 'residual' or 'fitted'.}

\item{var.method}{An integer (default: 1). The method to estimate the innovations variance for AR (See `ar.burg`)}

\item{ma.collector}{A character (default: "median"). A function name for collecting various MA models. (one of "mean", "median", or "pca)}

\item{return.mas}{A logical (default: TRUE). Whether it returns a data frame 'mas' containing various MAs.}

\item{elap.t}{A logical (default: FALSE). Whether it calculates elapsed time 'elap.t'.}

\item{verbose}{A logical (default: TRUE). Whether it message out about its processing.}
}
\value{
A list.
     `$x`: A original ts,
     `$out`: Final result ts,
     `$DD`: A list of intermediate result of difference,
     `$AR`: A list of intermediate result of AR,
     `$MA`: A ts of intermediate result of MA,
     `$EOA`: A ts of intermediate result of MAC,
     `$mas`: A data frame. Various MAs.
     `$NA.times`: A list with
         `$all` : time stamps whose values are NA,
         `$head`: time stamps whose values are NA at the head,
         `$tail`: time stamps whose values are NA at the tail,
     with attributes:
         `$d`: The optimal order of difference,
         `$p`: The optimal order of AR,
         `$q`: The order(s) of MA(C),
         `$ma.collector`: Used collector function for MAC,
         `$elap.t`: Elapsed time in second.
         `$suffix`: Final model name, e.g. "ARIMAC(10|1~10|1)"
}
\description{
(Difference -> AR -> MA -> band-pass)
}
