% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{shift.cyclic}
\alias{shift.cyclic}
\title{Cyclic shift a `ts`}
\usage{
shift.cyclic(ts, t_cyclic)
}
\arguments{
\item{ts}{A `ts` object.}

\item{t_cyclic}{A numeric. Time to be cyclic-shifted}
}
\value{
A cyclic-shifted `ts`.
}
\description{
Cyclic shift a `ts`
}
\examples{
# {not run}
# > xt <- ts(c(1,2,3,4,5), start=1, freqyency=1)
# > xt_cyc <- shift.cyclic(xt, t_cyclic=2)
# > xt_cyc
# Time Series:
# Start = 1
# End = 5
# Frequency = 1
# [1] 3 4 5 1 2

}
