% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{shift}
\alias{shift}
\title{Shift time series by time offset}
\usage{
shift(x, t_shift)
}
\arguments{
\item{x}{A `ts` object.}

\item{t_shift}{A numeric. Amount to shift (in seconds).}
}
\value{
A time-shifted `ts`.
}
\description{
Shift time series by time offset
}
