% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{shift_cyclic}
\alias{shift_cyclic}
\title{Cyclically shift a `ts`}
\usage{
shift_cyclic(x, t_cyclic)
}
\arguments{
\item{x}{A `ts` object.}

\item{t_cyclic}{A numeric. Time to shift cyclically.}
}
\value{
A cyclic-shifted `ts` object.
}
\description{
Cyclically shift a `ts`
}
