% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{shift_phase}
\alias{shift_phase}
\title{Align phase of a `ts` to another}
\usage{
shift_phase(ts, ref = NULL, phase = NULL)
}
\arguments{
\item{ts}{A `ts` object to be shifted.}

\item{ref}{A reference `ts` object (optional).}

\item{phase}{A numeric phase shift (optional).}
}
\value{
A phase-aligned `ts` object with attributes for applied shift.
}
\description{
Align phase of a `ts` to another
}
