% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{sigmasq}
\alias{sigmasq}
\title{Estimate signal loudness (sigma squared)}
\usage{
sigmasq(
  htilde,
  psd = NULL,
  low_frequency_cutoff = NULL,
  high_frequency_cutoff = NULL
)
}
\arguments{
\item{htilde}{A `ts` or `fs` object. The waveform.}

\item{psd}{An optional `fs` PSD object.}

\item{low_frequency_cutoff}{Lower frequency bound (Hz).}

\item{high_frequency_cutoff}{Upper frequency bound (Hz).}
}
\value{
Numeric. The unnormalized signal power.
}
\description{
Computes the matched-filter self-inner-product of a waveform in frequency domain.
}
\references{
https://pycbc.org/pycbc/latest/html/_modules/pycbc/filter/matchedfilter.html#sigmasq
}
