% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{sinking}
\alias{sinking}
\title{Save console output to a file}
\usage{
sinking(expr, file)
}
\arguments{
\item{expr}{An R expression to evaluate and capture output from.}

\item{file}{A character string. File path to save the console output.}
}
\value{
None. The console output of \code{expr} is written to \code{file}.
}
\description{
Redirects printed output of an expression to a file using \code{sink()}.
}
\examples{
\dontrun{
sinking(summary(cars), "summary.txt")
}

}
