% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{sprint}
\alias{sprint}
\title{Print Formatted String Using cat and sprintf}
\usage{
sprint(..., sep = "")
}
\arguments{
\item{...}{Arguments passed to \code{sprintf}.}

\item{sep}{A character string to separate values (default: "").}
}
\value{
None.
}
\description{
Print a string by combining \code{sprintf()} and \code{cat()}.
}
