% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{start_time}
\alias{start_time}
\title{Print start time or update start time}
\usage{
start_time(x, tstart = NULL)
}
\arguments{
\item{x}{A time series (`ts`) object.}

\item{tstart}{A numeric (default: NULL). A start time to be updated.}
}
\value{
Nothing, but given `ts.`'s start time will be updated.
}
\description{
Print start time or update start time
}
