% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{str.fs}
\alias{str.fs}
\title{Custom Structure Summary for `fs` Object}
\usage{
\method{str}{fs}(x)
}
\arguments{
\item{x}{An `fs` object.}
}
\value{
None. Prints structure to console.
}
\description{
Displays a concise structural summary of an `fs` object, highlighting class and delta_f.
}
