% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{sys_cat}
\alias{sys_cat}
\title{Print Contents of a File to Console}
\usage{
sys_cat(file)
}
\arguments{
\item{file}{A character string specifying the file path.}
}
\value{
None.
}
\description{
Print the contents of a file using system \code{cat}.
}
