% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{tf}
\alias{tf}
\title{Simple final time}
\usage{
tf(ts, n = 1)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{n}{A numeric (default: 1). A number of tail.}
}
\value{
Numeric(s). n final times.
}
\description{
Simple final time
}
