% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{ti}
\alias{ti}
\title{Simple initial time}
\usage{
ti(ts, n = 1)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{n}{A numeric (default: 1). A number of head.}
}
\value{
Numeric(s). n initial times.
}
\description{
Simple initial time
}
