% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{tl}
\alias{tl}
\title{Simple time length}
\usage{
tl(ts)
}
\arguments{
\item{ts}{A time series (`ts`) object.}
}
\value{
A vector of length 2 which is equivalent to `c(ti(ts), tf(ts))`.
}
\description{
Simple time length
}
