% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{to_fs}
\alias{to_fs}
\title{Transform `ts` class to `fs` class}
\usage{
to_fs(ts, delta_f = NULL)
}
\arguments{
\item{ts}{A `ts` object to be transformed to `fs`.}

\item{deltaf}{A numeric. A delta frequency.}
}
\value{
A `fs` object with respect to the given `ts`.
}
\description{
Transform `ts` class to `fs` class
}
