% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{to_ts}
\alias{to_ts}
\title{Transform `fs` class to `ts` class}
\usage{
to_ts(fs, start = 0, delta_t = NULL)
}
\arguments{
\item{fs}{A `fs` object}
}
\value{
A `ts` object with respect to the given `fs` object.
}
\description{
Transform `fs` class to `ts` class
}
