% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{tr}
\alias{tr}
\title{Simple time range}
\usage{
tr(ts)
}
\arguments{
\item{ts}{A time series (`ts`) object.}
}
\value{
A vector of length 2 which is equivalent to `c(ti(ts), tf(ts))`.
}
\description{
Simple time range
}
