% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{transpose.List}
\alias{transpose.List}
\title{Reshape a list-of-lists by transposing elements}
\usage{
transpose.List(list)
}
\arguments{
\item{list}{A list of named lists. Each sub-list must have the same names.}
}
\value{
A named list of lists, with outer and inner structure transposed.
}
\description{
This function reshapes a list of named lists into a named list of lists,
effectively performing a transpose operation. It is equivalent to `purrr::list_transpose()`,
but implemented in base R with minimal code.
}
\examples{
\dontrun{
x <- list(
    list(a = 1, b = 2),
    list(a = 3, b = 4)
)
transpose.List(x)
# Returns: list(a = list(1, 3), b = list(2, 4))
}
}
