% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detector.R
\name{travel.times}
\alias{travel.times}
\title{Light travel time among detectors}
\usage{
travel.times(dets = c("H1", "L1", "V1", "K1"))
}
\arguments{
\item{dets}{A vector. Charcters of detector abbreviations.}
}
\value{
A data frame. Light travel time values among detectors are in only upper triangular matrix.
}
\description{
Light travel time among detectors
}
