% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{ts.df}
\alias{ts.df}
\title{Transform to data frame with 'time' column}
\usage{
ts.df(ts, tzero = 0, val.name = "x")
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{tzero}{A numeric (default: 0). A zero time. Times will be shifted by '\code{tzero}'.}
}
\value{
A data frame with 'time' column at first.
}
\description{
Transform to data frame with 'time' column
}
