% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{ts_df}
\alias{ts_df}
\title{Convert `ts` to data frame with time column}
\usage{
ts_df(x, tzero = 0, val.name = "x")
}
\arguments{
\item{x}{A `ts` object.}

\item{tzero}{A numeric (default: 0). Time shift.}

\item{val.name}{A character. Name for value column (default: "x").}
}
\value{
A data frame with columns \code{time} and value.
}
\description{
Convert `ts` to data frame with time column
}
