% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{tsfy}
\alias{tsfy}
\title{Convert to `ts` object using a reference}
\usage{
tsfy(obj, ref, sampling.freq = NULL, na.rm = T)
}
\arguments{
\item{obj}{A vector, list, or data frame.}

\item{ref}{A `ts` object, time vector, or list specifying reference time.}

\item{sampling.freq}{Optional frequency if `ref` is numeric.}

\item{na.rm}{Logical. If \code{TRUE}, remove NA values (default: TRUE).}
}
\value{
A `ts` object or list of `ts` objects.
}
\description{
Convert to `ts` object using a reference
}
