% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{unit_normalize}
\alias{unit_normalize}
\title{Normalize a `ts` to unit scale}
\usage{
unit_normalize(x)
}
\arguments{
\item{x}{A `ts` object.}
}
\value{
A normalized `ts` object. Scale factor is stored in \code{attr(..., "order")}.
}
\description{
Normalize a `ts` to unit scale
}
