% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe.R
\name{update_stat}
\alias{update_stat}
\title{Update accumulated cluster-level statistics}
\usage{
update_stat(upd, cur)
}
\arguments{
\item{upd}{A list containing previous cumulative statistics (from `stat_anom()`), must contain `stats`.}

\item{cur}{A list containing statistics of the current batch (from `stat_anom()`), must contain `stats`.}
}
\value{
A list with a single element `stats`, containing:
\describe{
  \item{t_batch}{Cumulative batch duration.}
  \item{N_cl}{Cumulative number of detected clusters.}
  \item{N_anom}{Cumulative number of detected anomalies.}
  \item{lambda_c}{Updated cluster rate per second.}
  \item{lambda_a}{Updated average anomaly count per cluster.}
}
}
\description{
Updates the global cluster statistics by combining the previous accumulated statistics (`upd`)
and the current batch statistics (`cur`). The quantities updated include batch duration, number of clusters,
number of anomalies, and the derived rates \eqn{\lambda_c} (cluster rate) and \eqn{\lambda_a} (anomaly rate).
}
