% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{utc2gps}
\alias{utc2gps}
\title{Convert UTC time to GPS time}
\usage{
utc2gps(
  y,
  origin = as.Date("1980-01-06"),
  in_format = "\%Y-\%m-\%d \%H:\%M:\%OS"
)
}
\arguments{
\item{y}{A character. A UTC time.}

\item{origin}{A character (default: '1980-01-06'). A GPS time origin (Do not change for GW work).}

\item{format}{A character. A specific format of UTC time, e.g. `"%Y-%m-%d"`, `"%Y-%m-%d %H:%M:%OS"`.}
}
\value{
A numeric. A GPS time.
}
\description{
Convert UTC time to GPS time
}
