% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{vf}
\alias{vf}
\title{Simple final value extractor}
\usage{
vf(x, n = 1)
}
\arguments{
\item{x}{A vector.}

\item{n}{A numeric (default: 1). Number of final values to extract.}
}
\value{
A vector of the last n elements.
}
\description{
Simple final value extractor
}
