% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{vi}
\alias{vi}
\title{Simple initial value}
\usage{
vi(x, n = 1)
}
\arguments{
\item{x}{A vector.}

\item{n}{A numeric (default: 1). A number of head.}
}
\value{
Numeric(s). n initial times.
}
\description{
Simple initial value
}
