% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{vl}
\alias{vl}
\title{Length from first to last value}
\usage{
vl(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A numeric. Difference between last and first element.
}
\description{
Length from first to last value
}
