% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{weighted_inner}
\alias{weighted_inner}
\title{Weighted inner product}
\usage{
weighted_inner(x, y, wt)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{wt}{A weight vector.}
}
\value{
An inner product of `x`, `y` weighted by `wt`
}
\description{
This is not constrained as array.
`wt` needs to have also same dimension (or length) as x or y.
}
