% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{whiten}
\alias{whiten}
\title{Spectral Whitening}
\usage{
whiten(x, sl, fl, fu, ...)
}
\arguments{
\item{x}{A `ts` object.}

\item{sl}{A numeric. The segment length for estimating `psd`.}

\item{fl}{A numeric. The lower-frequency-cutoff for `bandpass`.}

\item{fu}{A numeric. The upper-frequency-cutoff for `bandpass`.}

\item{...}{Additional arguments for `bandpass()`.}
}
\value{
A `ts` object. Whitened time series.
}
\description{
Spectral Whitening
}
