% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{write.H5}
\alias{write.H5}
\title{Write Cleaned Data with a H5 file}
\usage{
write.H5(object, file, source, det)
}
\arguments{
\item{object}{A list. A result from "seqarima" (Not an object 'sequential.arima').}

\item{file}{A character. A file path to save.}

\item{source}{A character. An event name.}

\item{det}{A character. A detector name.}
}
\description{
Write Cleaned Data with a H5 file
}
