% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{write_H5}
\alias{write_H5}
\title{Write a time series and metadata into HDF5 format}
\usage{
write_H5(file, tsobj, meta.list = NULL)
}
\arguments{
\item{file}{Path to output HDF5 file.}

\item{tsobj}{A time series object of class `ts` (e.g., output of `ts()`).}

\item{meta.list}{A named list of additional metadata to store.
Names should be full HDF5 paths (e.g., "quality/simple/DQmask").}
}
\value{
Invisibly returns the HDF5 file path.
}
\description{
Write a time series and metadata into HDF5 format
}
