% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{amax}
\alias{amax}
\title{Calculate maximum of absolute values}
\usage{
amax(x, na.rm = T)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical (default: TRUE). Whether to remove NA values.}
}
\value{
A numeric. The maximum of absolute values.
}
\description{
Calculate maximum of absolute values
}
