% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{append.line}
\alias{append.line}
\title{Append line in txt file}
\usage{
append.line(x, file, show = T)
}
\arguments{
\item{x}{ANY.}

\item{file}{A character. An existing file path to append.}

\item{show}{A logical (default: TRUE). Whether it prints the file on console.}
}
\description{
Append line in txt file
}
