% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{apply_pca}
\alias{apply_pca}
\title{Apply PCA}
\usage{
apply_pca(x, retx = T, center = F, scale. = F, tol = NULL, rank. = NULL, ...)
}
\arguments{
\item{x}{A data frame or a matrix.}

\item{retx}{A logical (default: TRUE). See 'prcomp'.}

\item{center}{A logical (default: FALSE). See 'prcomp'.}

\item{scale.}{A logical (default: FALSE). See 'prcomp'.}

\item{tol}{See 'prcomp'.}

\item{rank.}{See 'prcomp'.}

\item{...}{Additional arguments for 'prcomp'.}
}
\value{
PCA object with fixed loadings' sign.
}
\description{
This function preserves the sign of loadings
}
