% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{as_tbt}
\alias{as_tbt}
\title{Transform to 'tbl_time'}
\usage{
as_tbt(ts, time_col = "time", val_col = "x")
}
\arguments{
\item{ts}{A time series (`ts`) object.}
}
\value{
A tbl_time.
}
\description{
Transform to 'tbl_time'
}
