% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{auto_diff}
\alias{auto_diff}
\title{Automatically determine differencing order via KPSS}
\usage{
auto_diff(ts, t.seg = 0.5, d_max = 2, verbose = TRUE)
}
\arguments{
\item{ts}{A `ts` object.}

\item{t.seg}{Segment duration.}

\item{d_max}{Maximum differencing order.}

\item{verbose}{Whether to print messages.}
}
\value{
A list with selected order, differenced output, and KPSS p-values.
}
\description{
Automatically determine differencing order via KPSS
}
