% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{bandpass}
\alias{bandpass}
\title{Band-pass filter}
\usage{
bandpass(
  ts,
  fl = NULL,
  fu = NULL,
  resp = "FIR",
  filt_order = NULL,
  verbose = T
)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{fl}{A numeric. The frequency lower bound for the band-pass filter.}

\item{fu}{A numeric. The frequency upper bound for the band-pass filter.}

\item{resp}{A character. "FIR (Finite Impulse Response)" or "IIR (Infinite Impulse Response)"}

\item{filt_order}{A numeric. The order of filter. Default value is 512 (FIR) or 8 (IIR).}
}
\value{
A band-pass filtered time series (`ts`) object.
}
\description{
Band-pass filter
}
