% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{burgar}
\alias{burgar}
\title{Fit AR model using modified Burg method}
\usage{
burgar(
  x,
  ic = "AIC",
  order.max = NULL,
  na.action = na.fail,
  demean = TRUE,
  series = NULL,
  var.method = 2L,
  numCores = NULL,
  ...
)
}
\arguments{
\item{x}{A `ts` object.}

\item{ic}{Information criterion: one of 'AIC', 'BIC', 'FPE', 'AICc', 'KIC', 'AKICc'.}

\item{order.max}{Maximum AR order.}

\item{na.action}{How to handle NA.}

\item{demean}{Whether to subtract the mean.}

\item{series}{Series name.}

\item{var.method}{Method for innovation variance.}

\item{...}{Additional args.}
}
\value{
An `ar` object.
}
\description{
Fit AR model using modified Burg method
}
\details{
For \code{ic}, followings can be chosen:
\itemize{
  \item AIC: Akaike Information Criterion (penalty factor: 2)
  \item BIC: Bayesian Information Criterion (penalty factor: log(n))
  \item FPE: Final Prediction Error
  \item AICc: Corrected AIC with bias correction for small samples
  \item KIC: Kashyap Information Criterion (penalty factor: 3)
  \item AKICc: Corrected KIC with bias correction
}
}
