% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{burgar}
\alias{burgar}
\title{Fit AR model using modified Burg method}
\usage{
burgar(
  x,
  ic = TRUE,
  order.max = NULL,
  na.action = na.fail,
  demean = TRUE,
  series = NULL,
  var.method = 2L,
  numCores = NULL,
  ...
)
}
\arguments{
\item{x}{A `ts` object.}

\item{ic}{Information criterion: logical or character.}

\item{order.max}{Maximum AR order.}

\item{na.action}{How to handle NA.}

\item{demean}{Whether to subtract the mean.}

\item{series}{Series name.}

\item{var.method}{Method for innovation variance.}

\item{...}{Additional args.}
}
\value{
An `ar` object.
}
\description{
Fit AR model using modified Burg method
}
