% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{check_normality}
\alias{check_normality}
\title{Check for normality using the Anderson-Darling test}
\usage{
check_normality(x, t.seg = 0.5)
}
\arguments{
\item{x}{A `ts` object.}

\item{t.seg}{Segment duration in seconds.}
}
\value{
A numeric vector of p-values.
}
\description{
Check for normality using the Anderson-Darling test
}
