% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{check_stationary}
\alias{check_stationary}
\title{Check for stationarity using the KPSS test across segments}
\usage{
check_stationary(x, t.seg = 0.5)
}
\arguments{
\item{x}{A `ts` object.}

\item{t.seg}{Segment duration in seconds.}
}
\value{
A matrix of KPSS p-values for level and trend.
}
\description{
Check for stationarity using the KPSS test across segments
}
