% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{correlate}
\alias{correlate}
\title{Correlate two complex vectors}
\usage{
correlate(x, y)
}
\arguments{
\item{x}{A complex vector.}

\item{y}{A complex vector.}
}
\value{
A complex vector, result of conjugate(x) * y.
}
\description{
Computes the element-wise product of the conjugate of `x` and `y`, used in matched filtering.
}
\references{
https://pycbc.org/pycbc/latest/html/_modules/pycbc/filter/matchedfilter_numpy.html#correlate
}
