% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{cutoff_to}
\alias{cutoff_to}
\title{Crop frequency series to a given range}
\usage{
cutoff_to(fs, ref)
}
\arguments{
\item{fs}{A `fs` object (frequency series).}

\item{ref}{A list with elements `min` and `max`, indicating index range.}
}
\value{
A cropped `fs` object with preserved attributes.
}
\description{
Crop frequency series to a given range
}
