% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{cyclic}
\alias{cyclic}
\title{Cyclically shift a vector}
\usage{
cyclic(x, n)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{An integer. Number of elements to shift.}
}
\value{
A cyclic-shifted vector.
}
\description{
Cyclically shift a vector
}
