% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTrasnform.R
\name{deltam_f}
\alias{deltam_f}
\title{Fractional mismatch between neighbouring tiles}
\usage{
deltam_f(mismatch)
}
\arguments{
\item{mismatch}{A numeric. percentage of desired fractional mismatch.}
}
\value{
A numeric.
}
\description{
Fractional mismatch between neighbouring tiles
}
