% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QT.R
\name{.iter_qs}
\alias{.iter_qs}
\title{Iterate over the Q values}
\usage{
.iter_qs(qrange, deltam)
}
\arguments{
\item{qrange}{A numeric vector of length 2. Lower and upper bounds of Q range.}

\item{deltam}{A numeric. Fractional mismatch between neighboring Q tiles.}
}
\value{
A numeric vector of Q values for Q-planes.
}
\description{
Compute Q values to use for Q-tiling, given a range and mismatch tolerance.
}
