% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{eoa}
\alias{eoa}
\title{Apply Ensemble of Averages (EoA)}
\usage{
eoa(ts, qs, collector = "median")
}
\arguments{
\item{ts}{A `ts` object.}

\item{qs}{MA orders.}

\item{collector}{Aggregation method.}
}
\value{
Smoothed `ts` object with MA ensemble metadata.
}
\description{
Apply Ensemble of Averages (EoA)
}
