% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{evenify}
\alias{evenify}
\title{Ensure even-length `ts`}
\usage{
evenify(x)
}
\arguments{
\item{x}{A `ts` object.}
}
\value{
A `ts` object with even length (truncates last sample if needed).
}
\description{
Ensure even-length `ts`
}
