% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{extract_pc}
\alias{extract_pc}
\title{Extract PC(s)}
\usage{
extract_pc(x, pc = "PC1")
}
\arguments{
\item{x}{A data frame or a matrix.}

\item{pc}{A character or a vector. Which PCs need to be extract? Naming convension: "PC#" (e.g. PC1, PC23, ...).}
}
\value{
A matrix or a vector.
}
\description{
Extract PC(s)
}
