% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{fs}
\alias{fs}
\title{Create a Frequency Series Object (`fs`)}
\usage{
fs(x, df, sampling.freq)
}
\arguments{
\item{x}{A numeric vector representing frequency-domain data.}

\item{df}{A numeric scalar. Frequency resolution (`delta_f`).}
}
\value{
An object of class `fs`, with additional attributes such as `delta_f` and `flen`.
}
\description{
Wraps a numeric vector into an `fs` class, assigning appropriate attributes
for frequency-domain analysis.
}
\examples{
fs(c(1, 2, 3), df = 0.5)
}
