% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pycbc.R
\name{get_antpatt}
\alias{get_antpatt}
\title{Calculate Antenna Pattern}
\usage{
get_antpatt(det, ra, dec, pol, t_gps, online = F)
}
\arguments{
\item{ra}{A numeric. RA of the GW source in radian.}

\item{dec}{A numeric. Dec of the GW source in radian.}

\item{pol}{A numeric. Polarization angle of the GW source in radian.}

\item{t_gps}{A numeric. Passing time of the GW source to Earth in gps second.}

\item{dets}{A vector. Characters of detector abbreviations.}
}
\value{
A list. It has two elements of factors for the plus (+) polarization and the cross (x) polarization.
}
\description{
Calculate Antenna Pattern
}
