% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pycbc.R
\name{get_available_detectors}
\alias{get_available_detectors}
\title{Get available GW detectors}
\usage{
get_available_detectors()
}
\value{
A data frame with columns: \code{Detector Name}, \code{Abbreviation}, \code{latitude}, \code{longitude}.
}
\description{
Queries the list of available detectors from PyCBC and returns location info.
}
\details{
This function interfaces with Python's \pkg{pycbc} library using \pkg{reticulate}.
Please ensure that the required Python packages are available in your environment.
}
