% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSC.R
\name{get_gwosc}
\alias{get_gwosc}
\title{Retrieve metadata from GWOSC}
\usage{
get_gwosc(offline = F, csvpath = NULL)
}
\arguments{
\item{offline}{Logical. If TRUE, read metadata from a local CSV file instead of querying online.}

\item{csvpath}{Character. Required if `offline = TRUE`; path to the local CSV file.}
}
\value{
A data.frame containing metadata of events.
}
\description{
Download or load offline metadata table of gravitational wave events from GWOSC.
}
