% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSC.R
\name{get_gwosc_param}
\alias{get_gwosc_param}
\title{Extract specific parameters from GWOSC metadata}
\usage{
get_gwosc_param(gwosc.list, source.names, param)
}
\arguments{
\item{gwosc.list}{A data.frame returned by `get_gwosc()`.}

\item{source.names}{Character vector. Names of the events (`commonName`) to extract.}

\item{param}{Character or character vector. Column name(s) to extract.}
}
\value{
A vector or data.frame depending on the number of requested parameters.
}
\description{
Extract specific parameters from GWOSC metadata
}
