% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSC.R
\name{get_gwosc_param}
\alias{get_gwosc_param}
\title{Get GWTC parameter from the return of get.gwosc()}
\usage{
get_gwosc_param(gwosc.list, source.names, param)
}
\arguments{
\item{source.names}{A character vector.
Source names need to follow the commonName convention.
It can be either one character or a vector of source.names.}

\item{param}{A character vector.
Parameters that want to get.
It can be either one character or a vector of source.names.}

\item{gwossc.list}{A data.frame. The return object from get.gwosc()}
}
\description{
Get GWTC parameter from the return of get.gwosc()
}
