% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{get_MissingValues}
\alias{get_MissingValues}
\title{Sequential ARIMA pipeline components}
\usage{
get_MissingValues(x, ref)
}
\arguments{
\item{x}{A time series (`ts`) object that may have NA values.}

\item{ref}{A reference `ts` object used for comparison.}
}
\value{
A list with:
\describe{
  \item{all}{All time stamps with NA in `ts` compared to `ref`}
  \item{head}{NA timestamps at the start of the series}
  \item{tail}{NA timestamps at the end of the series}
}
}
\description{
This file contains all key components used in the seqarima pipeline,
including differencing, autoregressive (AR) fitting, and moving average (MA) smoothing functions.
}
\details{
Identify missing values in a time series relative to a reference
}
