% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pycbc.R
\name{get_wave}
\alias{get_wave}
\title{Loading waveforms from PyCBC}
\usage{
get_wave(
  model.name,
  sampling.freq,
  fl = NULL,
  fu = NULL,
  m1 = NULL,
  m2 = NULL,
  d_L = NULL,
  inc = 0,
  ...,
  R.ts = TRUE,
  det,
  ra,
  dec,
  pol,
  t_gps,
  proj = FALSE
)
}
\description{
Loading waveforms from PyCBC
}
