% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{gps2utc}
\alias{gps2utc}
\title{Convert GPS time to UTC time}
\usage{
gps2utc(
  x,
  origin = as.Date("1980-01-06"),
  out_format = "\%Y-\%m-\%d \%H:\%M:\%OS"
)
}
\arguments{
\item{x}{A numeric. A GPS time.}

\item{origin}{A character (default: '1980-01-06'). A GPS time origin (Do not change for GW work).}

\item{formatting}{A logical (default: FALSE). Whether it formats with "%Y-%m-%d %H:%M:%OS", otherwise, "YYYYMMDD HH:MM:SS".}
}
\value{
A character. An UTC time.
}
\description{
Convert GPS time to UTC time
}
