% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{gps2utc}
\alias{gps2utc}
\title{Convert GPS time to UTC}
\usage{
gps2utc(x, origin = as.POSIXct("1980-01-06 00:00:00", tz = "UTC"))
}
\arguments{
\item{x}{A numeric. GPS time.}

\item{origin}{A POSIXct (default: as.POSIXct("1980-01-06 00:00:00", tz = "UTC"))}
}
\value{
A POSIXct of UTC time.
}
\description{
Convert GPS time to UTC
}
