% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{inner}
\alias{inner}
\title{Inner product}
\usage{
inner(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
An inner product of `x`, `y`.
}
\description{
This is not constrained as array.
}
