% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{interp_psd}
\alias{interp_psd}
\title{Interpolate PSD
Return a new PSD that has been interpolated to the desired delta_f}
\usage{
interp_psd(x, delta_f)
}
\arguments{
\item{x}{A `fs` object. Frequency series to be interpolated.}

\item{delta_f}{A numeric. The desired delta_f of the output}
}
\description{
Interpolate PSD
Return a new PSD that has been interpolated to the desired delta_f
}
