% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{inv_spec_trunc_psd}
\alias{inv_spec_trunc_psd}
\title{Inverse Spectrum Truncation}
\usage{
inv_spec_trunc_psd(x, max_filter_len, fl = NULL, trunc_method = NULL)
}
\arguments{
\item{x}{A `fs` object of PSD.}

\item{max_filter_len}{A numeric.}

\item{fl}{A numeric. Low-frequency-cutoff.}

\item{trunc_method}{A window function. For truncating spectrum.}
}
\description{
Inverse Spectrum Truncation
}
