% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{ma}
\alias{ma}
\title{Two-sided centered Moving Average smoother}
\usage{
ma(ts, order, na.rm = T)
}
\arguments{
\item{ts}{A `ts` object.}

\item{order}{MA order.}

\item{na.rm}{Remove NAs.}
}
\value{
Smoothed `ts` object.
}
\description{
Two-sided centered Moving Average smoother
}
