% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{message_by}
\alias{message_by}
\title{Print Message Conditionally by Iteration}
\usage{
message_by(i = i, verbose = T, msg)
}
\arguments{
\item{i}{Integer indicating current iteration index.}

\item{verbose}{Logical or integer; if \code{TRUE}, print every iteration; if integer, print every \code{verbose}-th iteration.}

\item{msg}{A character string to be printed.}
}
\value{
None.
}
\description{
Print a message every \code{verbose}-th iteration in a loop.
}
