% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{mmed}
\alias{mmed}
\title{Apply median filter to a `ts`}
\usage{
mmed(ts, order)
}
\arguments{
\item{ts}{A `ts` object.}

\item{order}{An odd integer. Window size of the median filter.}
}
\value{
A smoothed `ts` object.
}
\description{
Apply median filter to a `ts`
}
