% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{pad}
\alias{pad}
\title{Pad a `ts` with zeros}
\usage{
pad(ts, tstart, tend, at = 0)
}
\arguments{
\item{ts}{A `ts` object.}

\item{tstart}{A numeric. Start time of output.}

\item{tend}{A numeric. End time of output.}

\item{at}{A numeric. Time to inject original `ts`.}
}
\value{
A padded `ts` object.
}
\description{
Pad a `ts` with zeros
}
