% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{pad}
\alias{pad}
\title{Padding short ts into longer zeros.}
\usage{
pad(ts, tstart, tend, at = 0)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{tstart}{A numeric. Start time of zeros.}

\item{tend}{A numeric. End time of zeros.}

\item{at}{A numeric (default: 0). Time that ts will be injected.}
}
\value{
A ts object of given ts injected zeros.
}
\description{
Padding short ts into longer zeros.
}
