% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{paste_wd}
\alias{paste_wd}
\title{Paste WD to file name}
\usage{
paste_wd(file, prefix = NULL, dir = getwd())
}
\arguments{
\item{file}{A character. File name.}

\item{prefix}{A character. '{prefix}_file' will be the final file name.}

\item{dir}{A character (default: getwd()). Working directory will be pasted.}
}
\value{
A character.
}
\description{
Paste WD to file name
}
