% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{plot.fs}
\alias{plot.fs}
\title{Plot a Frequency Series (`fs`)}
\usage{
\method{plot}{fs}(
  x,
  ...,
  log = "xy",
  xlab = "Frequency (Hz)",
  ylab = "PSD",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{x}{An `fs` object.}

\item{...}{Additional arguments in `geom_line()`.}

\item{log}{A character ("x", "y", or "xy") to control log scaling.}

\item{xlab}{A character. X-axis label (default: "Frequency (Hz)").}

\item{ylab}{A character. Y-axis label (default: "PSD").}

\item{xlim}{A numeric vector of length 2 specifying x-axis limits.}

\item{ylim}{A numeric vector of length 2 specifying y-axis limits.}
}
\value{
A ggplot object.
}
\description{
Visualizes the amplitude (or PSD) of an `fs` object using ggplot2.
}
