% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{plot_oscillo}
\alias{plot_oscillo}
\title{Plot a Single Oscillogram}
\usage{
plot_oscillo(
  x,
  tzero = NULL,
  trange = NULL,
  ylim = "pm",
  title = NULL,
  size = 0.3,
  ...
)
}
\arguments{
\item{x}{A numeric time series object.}

\item{tzero}{A numeric value indicating time shift (default: NULL).}

\item{trange}{A numeric vector specifying time range to display.}

\item{ylim}{Either 'pm' or a numeric vector of length 2.}

\item{title}{A character string specifying the plot title.}

\item{size}{A numeric value for line width (default: 0.3).}

\item{...}{Additional arguments for \code{ggplot2::geom_line}.}
}
\value{
A ggplot object.
}
\description{
Generate a line plot of a single time series (oscillogram).
}
