% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{plot_spectro}
\alias{plot_spectro}
\title{Plot spectrogram}
\usage{
plot_spectro(
  ts,
  tzero = 0,
  trange = NULL,
  frange = c(32, 512),
  qrange = c(40, 1),
  crange = NULL,
  tres = 1000,
  fres = 1000,
  logf = T,
  title = NULL,
  specScale = FALSE,
  specXlabel = FALSE,
  specYlabel = TRUE,
  specGrid = "none",
  specColorPal = viridis::viridis(256),
  trans = NULL,
  specScaleDir = "vertical",
  specScalePos = "ul",
  linecolor = "black",
  osciLegend = FALSE,
  osciXlabel = TRUE,
  osciYlabel = TRUE,
  stack = T,
  ...
)
}
\arguments{
\item{ts}{A time series (\code{ts}) object.}

\item{tzero}{A numeric (default: \code{0}). A zero time. Times will be shifted by \code{tzero}.}

\item{trange}{A vector (default: \code{tr(ts)}). A time range for x axis limit.}

\item{frange}{A vector (default: \code{c(32, 512)}). A frequency range of y axis limit.}

\item{qrange}{A vector (default: \code{c(40, 1)}). A q range as an input to \code{qtransform} function.}

\item{crange}{A vector (default: \code{NULL}). A color range. A z axis range.}

\item{tres}{A numeric (default: \code{0.001}). A time resolution.}

\item{fres}{A numeric (default: \code{1000}). A frequency resolution.}

\item{logf}{A logical (dafault: \code{TRUE}). Whether transform frequency axis to log (base: 10) scale.}

\item{title}{A character. A title for the plot.}

\item{specScale}{A logical (default: \code{FALSE}). Whether it shows color scale bar for spectrogram.}

\item{specXlabel}{A logical (default: \code{FALSE}). Whether it shows x axis label for spectrogram.}

\item{specYlabel}{A logical (default: \code{TRUE}). Whether it shows y axis label for spectrogram.}

\item{specGrid}{A character (default: \code{"none"}). Overlay grid on x- ("x"), y-axis ("y"), or both ("xy").}

\item{specColorPal}{A Palette for spectrogram (default: \code{viridis::viridis(256)}).}

\item{trans}{A function (defualt: \code{NULL}). If function is given, spectrogram will be transformed by a function of \code{trans} (e.g. \code{trans = sqrt} for square-root).}

\item{specScaleDir}{A character (default: \code{"vertical"}). If \code{specScale=TRUE}, the direction of color bar. (\code{"vertical"} or \code{"horizontal"})}

\item{specScalePos}{A character (default: \code{"ul"}). If \code{specScale=TRUE}, the position of color bar. (\code{"ul"}, \code{"ur"}, \code{"bl"}, or \code{"br"})}

\item{osciLegend}{A logical (default: \code{FALSE}). Whether it shows a legend for oscillogram.}

\item{osciXlabel}{A logical (default: \code{FALSE}). Whether it shows x axis label for oscillogram.}

\item{osciYlabel}{A logical (default: \code{TRUE}). Whether it shows y axis label for oscillogram.}

\item{stack}{A logical (default: \code{TRUE}). Whether it stack spectrogram and oscillogram.}

\item{...}{Additional arguments for \code{ggpubr::ggarrange}.}
}
\value{
A ggplot object. If \code{stack=FALSE}, A list with \code{$spec.plot}: spectrogram ggplot object, \code{$osci.plot}: oscillogram ggplot object.
}
\description{
Using ggplot2 to generate q-transformed spectrogram & oscillogram together!
}
