% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detector.R
\name{proj_wave}
\alias{proj_wave}
\title{The projection process can also take into account the rotation of the Earth using the project wave function.}
\usage{
proj_wave(hp, hc, det, ra, dec, pol, t_gps)
}
\arguments{
\item{hp}{A ts object. A plus polarization signal template.}

\item{hc}{A ts object. A cross polarization signal template.}

\item{ra}{A numeric. RA of the GW source in radian.}

\item{dec}{A numeric. Dec of the GW source in radian.}

\item{pol}{A numeric. Polarization angle of the GW source in radian.}

\item{t_gps}{A numeric. Passing time of the GW source to Earth in gps second.}

\item{dets}{A vector. Characters of detector abbreviations.}
}
\value{
A Python pycbc object. A projected signal template.
}
\description{
The projection process can also take into account the rotation of the Earth using the project wave function.
}
