% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{psd}
\alias{psd}
\title{Generate PSD}
\usage{
psd(ts, sl = 4, fl = 15, delf = NULL, window_func = bspec::hannwindow)
}
\arguments{
\item{ts}{A `ts` object.}

\item{sl}{A numeric. A segment length.}

\item{fl}{A numeric. Low-frequency-cutoff.}
}
\value{
A `fs` object. PSD with interpolated and inverse-spectrum-truncated.
}
\description{
Generate PSD
}
