% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QT.R
\name{qplane}
\alias{qplane}
\title{Compute Q-plane across all Q and f0 tiles}
\usage{
qplane(qplane.tile.list, fseries, return_complex = FALSE)
}
\arguments{
\item{qplane.tile.list}{A list of Q-tiles (from \code{qtiling}).}

\item{fseries}{An \code{fs} object. Frequency-series representation of the data.}

\item{return_complex}{Logical. If \code{TRUE}, return complex-valued transform.}
}
\value{
A list with:
\describe{
  \item{times}{Time axis (s)}
  \item{freqs}{Frequency axis (Hz)}
  \item{plane}{Q-transform output (matrix)}
}
}
\description{
Evaluate all Q-tiles and construct the Q-transform energy (or complex) plane.
}
