% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QT.R
\name{qseries}
\alias{qseries}
\title{Q-series}
\usage{
qseries(fseries, Q, f0, return_complex = FALSE)
}
\arguments{
\item{fseries}{A `fs` object. frequency-series data set.}

\item{Q}{A numeric. q value.}

\item{f0}{A numeric. central frequency.}

\item{return_complex}{A logical (default: FALSE). Return the raw complex series instead of the normalized power.}
}
\value{
A `ts` of the normalized energy from the Q-transform of this tile against the data.
}
\description{
Q-series
}
