% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTrasnform.R
\name{qtiling}
\alias{qtiling}
\title{Q-tiling}
\usage{
qtiling(fseries, qrange, frange, mismatch = 0.2)
}
\arguments{
\item{fseries}{A `fs` object. frequency-series data set.}

\item{qrange}{A numeric vector. upper and lower bounds of q range.}

\item{frange}{A numeric vector. upper and lower bounds of frequency range.}

\item{mismatch}{A numeric. percentage of desired fractional mismatch.}
}
\value{
A list containing Q-tile vectors for a set of Q-planes.
}
\description{
Q-tiling
}
