% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{reinstall.beacon}
\alias{reinstall.beacon}
\title{Reinstall beacon}
\usage{
reinstall.beacon(beacon.path = "~/projects/GW/beacon")
}
\arguments{
\item{beacon.path}{A character. A beacon package path.}
}
\value{
reloading
}
\description{
Reinstall beacon
}
