% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{resize}
\alias{resize}
\title{Resize a `ts` to target length}
\usage{
resize(ts, nlen)
}
\arguments{
\item{ts}{A `ts` object.}

\item{nlen}{An integer. Desired length.}
}
\value{
A resized `ts` object (padded with zeros or truncated).
}
\description{
Resize a `ts` to target length
}
