% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{savemsg}
\alias{savemsg}
\title{Save a Message to a Text File}
\usage{
savemsg(msg, ..., file, show = T)
}
\arguments{
\item{msg}{A character string representing the message to save.}

\item{...}{Additional character strings to be saved.}

\item{file}{A character string specifying the file path.}

\item{show}{Logical; if \code{TRUE}, print the message on console.}
}
\value{
None.
}
\description{
Append a message (or multiple messages) to a text file.
}
