% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{savetab}
\alias{savetab}
\title{Save a Data Frame as a Markdown-Formatted Table}
\usage{
savetab(
  x,
  file,
  signif = 3,
  col.names = F,
  row.names = F,
  quote = F,
  show = T,
  append = F,
  nsep = 30,
  ...
)
}
\arguments{
\item{x}{A data.frame to save.}

\item{file}{A character string indicating the file path.}

\item{signif}{Integer; number of significant digits (default: 3).}

\item{col.names}{Logical; whether to include column names (default: FALSE).}

\item{row.names}{Logical; whether to include row names (default: FALSE).}

\item{quote}{Logical; whether to quote each value (default: FALSE).}

\item{show}{Logical; if \code{TRUE}, print the table content (default: TRUE).}

\item{append}{Logical; if \code{TRUE}, append to the file (default: FALSE).}

\item{nsep}{Integer; number of dashes used to separate appended tables (default: 30).}

\item{...}{Additional arguments for \code{insight::export_table}.}
}
\value{
None.
}
\description{
Save a data frame in markdown format using \code{insight::export_table}.
}
