% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqARIMA.R
\name{seqarima}
\alias{seqarima}
\title{Sequential ARIMA function}
\usage{
seqarima(
  ts,
  d = NULL,
  p = NULL,
  q = NULL,
  fl = NULL,
  fu = NULL,
  ar.aic = TRUE,
  ar.collector = "median",
  ma.collector = "median",
  return.every = FALSE,
  elap.t = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{ts}{A time series (`ts`) object.}

\item{d}{An integer. The order of difference.}

\item{p}{An integer. The order of AR.}

\item{q}{An integer. The order of MA.}

\item{fl}{A numeric. The frequency lower bound for the band-pass filter.}

\item{fu}{A numeric. The frequency upper bound for the band-pass filter.}

\item{ar.aic}{A logical (default: TRUE). Whether p is selected by AIC or fixed.}

\item{ar.collector}{A character (default: "median"). A function name for collecting various AR residuals. (one of "mean", "median", or "pca")}

\item{ma.collector}{A character (default: "median"). A function name for collecting various MA models. (one of "mean", "median", or "pca")}

\item{return.every}{A logical (default: FALSE). Whether returns ts of every stage.}

\item{elap.t}{A logical (default: FALSE). Whether it calculates elapsed time 'elap.t'.}

\item{verbose}{A logical (default: TRUE). Whether it message out about its processing.}
}
\value{
A list.
     `$x`: A original ts,
     `$out`: Final result ts,
     `$DD`: A list of intermediate result of difference,
     `$AR`: A list of intermediate result of AR,
     `$MA`: A ts of intermediate result of MA,
     `$EOA`: A ts of intermediate result of EoA,
     `$NA.times`: A list with
         `$all` : time stamps whose values are NA,
         `$head`: time stamps whose values are NA at the head,
         `$tail`: time stamps whose values are NA at the tail,
     with attributes:
         `$d`: The optimal order of difference,
         `$p`: The optimal order of AR,
         `$q`: The order(s) of MA(C),
         `$ar.collector`: Used collector function for EAR,
         `$ma.collector`: Used collector function for EoA,
         `$elap.t`: Elapsed time in second.
         `$suffix`: Final model name, e.g. "ARIMAC(10|1~10|1)"
}
\description{
(Difference -> AR -> MA -> band-pass)
}
