% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{shift}
\alias{shift}
\title{Shift time series by time}
\usage{
shift(ts, t_shift)
}
\arguments{
\item{ts}{A `ts` object.}

\item{t_shift}{A numeric. Time to be shifted.}
}
\value{
A time-shifted `ts`.
}
\description{
Shift time series by time
}
\examples{
# {not run}
# > xt <- ts(c(1,2,3), start=1, frequency=1)
# > ti(xt) # >>> 1
# > xt_shift <- shift(xt, 1)
# > ti(xt_shift) # >>> 2

}
