% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoC.R
\name{shift_phase}
\alias{shift_phase}
\title{Shift phase of ts}
\usage{
shift_phase(ts, ref = NULL, phase = NULL)
}
\arguments{
\item{ts}{A ts object to be phase-shifted}

\item{ref}{A ts object as a reference (default: NULL).}

\item{phase}{A numeric (default: NULL). A specific shifting phase.}
}
\value{
A phase-shifted phase.
}
\description{
Shift phase of ts
}
