% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{sigmasq}
\alias{sigmasq}
\title{Calculate simga square}
\usage{
sigmasq(
  htilde,
  psd = NULL,
  low_frequency_cutoff = NULL,
  high_frequency_cutoff = NULL
)
}
\description{
Return the loudness of the waveform. This is defined (see Duncan Brown's
thesis) as the unnormalized matched-filter of the input waveform, htilde,
with itself. This quantity is usually referred to as (sigma)^2 and is then
used to normalize matched-filters with the data.
}
