% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{start_time}
\alias{start_time}
\title{Get or update the start time of a time series}
\usage{
start_time(x, tstart = NULL)
}
\arguments{
\item{x}{A `ts` object.}

\item{tstart}{A numeric. If provided, updates start time.}
}
\value{
If \code{tstart} is \code{NULL}, returns initial time. Otherwise, updates the object in global environment.
}
\description{
Get or update the start time of a time series
}
