% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{tf}
\alias{tf}
\title{Final timestamp of a time series}
\usage{
tf(x, n = 1)
}
\arguments{
\item{x}{A `ts` object.}

\item{n}{A numeric (default: 1). Number of final timestamps to return.}
}
\value{
A numeric vector of length `n`.
}
\description{
Final timestamp of a time series
}
