% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{ti}
\alias{ti}
\title{Initial timestamp of a time series}
\usage{
ti(x, n = 1)
}
\arguments{
\item{x}{A `ts` object.}

\item{n}{A numeric (default: 1). Number of initial timestamps to return.}
}
\value{
A numeric vector of length `n`.
}
\description{
Initial timestamp of a time series
}
