% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{to_fs}
\alias{to_fs}
\title{Convert Time Series (`ts`) to Frequency Series (`fs`)}
\usage{
to_fs(ts, delta_f = NULL)
}
\arguments{
\item{ts}{A `ts` object.}

\item{delta_f}{Optional. A numeric value indicating desired frequency resolution.
If NULL, defaults to `sampling.freq / length(ts)`.}
}
\value{
An `fs` object containing frequency-domain representation and metadata.
}
\description{
Transforms a time series object to a frequency series by computing FFT.
}
