% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{tr}
\alias{tr}
\title{Time range of a time series}
\usage{
tr(ts)
}
\arguments{
\item{ts}{A `ts` object.}
}
\value{
A numeric vector of length 2: \code{c(start, end)}.
}
\description{
Time range of a time series
}
