% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{tsfy}
\alias{tsfy}
\title{Transform to ts obj with reference ts}
\usage{
tsfy(obj, ref, sampling.freq = NULL, na.rm = T)
}
\arguments{
\item{obj}{A list or a vector.}

\item{ref}{A reference time series.}

\item{na.rm}{A logical (default: TRUE). Whether remove NA values or not.}
}
\value{
A time series (`ts`) object with same time stamp of 'ref'.
}
\description{
Transform to ts obj with reference ts
}
