% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{utc2gps}
\alias{utc2gps}
\title{Convert UTC to GPS time}
\usage{
utc2gps(
  y,
  origin = as.Date("1980-01-06"),
  in_format = "\%Y-\%m-\%d \%H:\%M:\%OS"
)
}
\arguments{
\item{y}{A character. UTC time string.}

\item{origin}{A character or Date (default: "1980-01-06"). GPS epoch.}

\item{in_format}{A character. Format string of the UTC time.}
}
\value{
A numeric. GPS time.
}
\description{
Convert UTC to GPS time
}
