% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{utc2gps}
\alias{utc2gps}
\title{Convert UTC to GPS time}
\usage{
utc2gps(y, origin = as.POSIXct("1980-01-06 00:00:00", tz = "UTC"))
}
\arguments{
\item{y}{A character. UTC time string.}

\item{origin}{A POSIXct (default: as.POSIXct("1980-01-06 00:00:00", tz = "UTC"))}
}
\value{
A numeric. GPS time.
}
\description{
Convert UTC to GPS time
}
