% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{vf}
\alias{vf}
\title{Simple final value}
\usage{
vf(x, n = 1)
}
\arguments{
\item{x}{A vector.}

\item{n}{A numeric (default: 1). A number of tail.}
}
\value{
Numeric(s). n final times.
}
\description{
Simple final value
}
