% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{vi}
\alias{vi}
\title{Simple initial value extractor}
\usage{
vi(x, n = 1)
}
\arguments{
\item{x}{A vector.}

\item{n}{A numeric (default: 1). Number of initial values to extract.}
}
\value{
A vector of the first n elements.
}
\description{
Simple initial value extractor
}
