% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{vl}
\alias{vl}
\title{Simple value length from first data point to last.}
\usage{
vl(x)
}
\arguments{
\item{ts}{A time series (`ts`) object.}
}
\value{
A vector of length 2 which is equivalent to `c(ti(ts), tf(ts))`.
}
\description{
Simple value length from first data point to last.
}
