% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\name{weighted_inner}
\alias{weighted_inner}
\title{Weighted inner product}
\usage{
weighted_inner(x, y, wt)
}
\arguments{
\item{x}{A numeric or complex vector.}

\item{y}{A numeric or complex vector.}

\item{wt}{A numeric weight vector (same length as `x`, `y`).}
}
\value{
Complex scalar. Weighted inner product.
}
\description{
Computes the inner product between `x` and `y` with inverse-variance weighting.
}
