% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{window_to}
\alias{window_to}
\title{Handy window function}
\usage{
window_to(ts, ref)
}
\arguments{
\item{ref}{A time series (`ts`) or a vector containing both an initial time and a final time.}

\item{x}{A time series (`ts`) object.}
}
\value{
A windowed time series (`ts`) object.
}
\description{
Handy window function
}
